### 2024-01-10 #6204 @editorialbot
[PRE REVIEW]: NOMAD CAMELS: Configurable Application for Measurements, Experiments and Laboratory Systems
**Submitting author:** @lapmk (<a href="http://orcid.org/0000-0003-1480-9161">Michael Krieger</a>)
**Repository:** https://github.com/FAU-LAP/NOMAD-CAMELS
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v0.2.3
