### 2024-01-10 #6204 @kyleniemeyer
Hello @lapmk, thanks for your submission to JOSS. We'll use this pre-review issue to assign an editor and find reviewers; unfortunately, we don't have any editors available to handle your submission right now, so I need to put this on our waitlist until someone can edit it.

In the meantime, any reviewer recommendations you can make would be welcome.
