### 2024-01-12 ed0ef02
Add extension functionality and manage extensions action, not yet working!
