+++ nomad_camels/gui/mainWindow_v2.py
@@ -81,6 +81,8 @@ class Ui_MainWindow(object):
         self.actionEPICS_driver_builder.setObjectName(u"actionEPICS_driver_builder")
         self.actionExport_from_databroker = QAction(MainWindow)
         self.actionExport_from_databroker.setObjectName(u"actionExport_from_databroker")
+        self.actionManage_Extensions = QAction(MainWindow)
+        self.actionManage_Extensions.setObjectName(u"actionManage_Extensions")
         self.centralwidget = QWidget(MainWindow)
         self.centralwidget.setObjectName(u"centralwidget")
         self.gridLayout_5 = QGridLayout(self.centralwidget)
@@ -138,13 +140,13 @@ class Ui_MainWindow(object):
 
         self.gridLayout_5.addWidget(self.pushButton_clear_log, 8, 9, 1, 2)
 
-        self.widget = QWidget(self.centralwidget)
-        self.widget.setObjectName(u"widget")
-        self.widget.setMaximumSize(QSize(16777215, 50))
-        self.gridLayout_6 = QGridLayout(self.widget)
+        self.menu_widget = QWidget(self.centralwidget)
+        self.menu_widget.setObjectName(u"menu_widget")
+        self.menu_widget.setMaximumSize(QSize(16777215, 50))
+        self.gridLayout_6 = QGridLayout(self.menu_widget)
         self.gridLayout_6.setObjectName(u"gridLayout_6")
         self.gridLayout_6.setContentsMargins(0, 0, 0, 0)
-        self.pushButton_manage_instr = QPushButton(self.widget)
+        self.pushButton_manage_instr = QPushButton(self.menu_widget)
         self.pushButton_manage_instr.setObjectName(u"pushButton_manage_instr")
         self.pushButton_manage_instr.setMaximumSize(QSize(150, 16777215))
         font = QFont()
@@ -154,32 +156,32 @@ class Ui_MainWindow(object):
 
         self.gridLayout_6.addWidget(self.pushButton_manage_instr, 0, 0, 1, 1)
 
-        self.line_2 = QFrame(self.widget)
+        self.line_2 = QFrame(self.menu_widget)
         self.line_2.setObjectName(u"line_2")
         self.line_2.setFrameShape(QFrame.VLine)
         self.line_2.setFrameShadow(QFrame.Sunken)
 
         self.gridLayout_6.addWidget(self.line_2, 0, 13, 1, 1)
 
-        self.line = QFrame(self.widget)
+        self.line = QFrame(self.menu_widget)
         self.line.setObjectName(u"line")
         self.line.setFrameShape(QFrame.VLine)
         self.line.setFrameShadow(QFrame.Sunken)
 
         self.gridLayout_6.addWidget(self.line, 0, 8, 1, 1)
 
-        self.comboBox_user_type = QComboBox(self.widget)
+        self.comboBox_user_type = QComboBox(self.menu_widget)
         self.comboBox_user_type.setObjectName(u"comboBox_user_type")
         self.comboBox_user_type.setMaximumSize(QSize(110, 16777215))
 
         self.gridLayout_6.addWidget(self.comboBox_user_type, 0, 1, 1, 1)
 
-        self.widget_4 = QWidget(self.widget)
-        self.widget_4.setObjectName(u"widget_4")
-        self.gridLayout_7 = QGridLayout(self.widget_4)
+        self.sample_widget = QWidget(self.menu_widget)
+        self.sample_widget.setObjectName(u"sample_widget")
+        self.gridLayout_7 = QGridLayout(self.sample_widget)
         self.gridLayout_7.setObjectName(u"gridLayout_7")
[…]
