### 2024-01-16 0dafb7c
Update NOMAD-CAMELS to NOMAD CAMELS in UI and configuration files, switched order in window titles, first the title, then NOMAD CAMELS
