+++ nomad_camels/ui_widgets/warn_popup.py
@@ -5,7 +5,7 @@ class WarnPopup(QMessageBox):
     def __init__(self, parent=None, text='', title='', info_icon=False):
         super().__init__(parent=parent)
         self.setText(text)
-        self.setWindowTitle(f'{title} - NOMAD-CAMELS')
+        self.setWindowTitle(f'{title} - NOMAD CAMELS')
         print(text)
         if info_icon:
             self.setIcon(QMessageBox.Information)

[…]
