### 2024-01-17 cf8978d
Add exception handling in device instantiation
