### 2024-01-17 5076754
Fix issue of stage control not updating set_channel when moving manually
