+++ nomad_camels/gui/plot_definer.py
@@ -22,19 +22,7 @@ from PySide6.QtWidgets import (QApplication, QCheckBox, QFrame, QGridLayout,
 from nomad_camels.ui_widgets.variable_tool_tip_box import Variable_Box
 
 class Ui_Plot_Definer(object):
-    """ """
     def setupUi(self, Plot_Definer):
-        """
-
-        Parameters
-        ----------
-        Plot_Definer :
-            
-
-        Returns
-        -------
-
-        """
         if not Plot_Definer.objectName():
             Plot_Definer.setObjectName(u"Plot_Definer")
         Plot_Definer.resize(470, 350)
@@ -53,12 +41,6 @@ class Ui_Plot_Definer(object):
         self.label_5.setFont(font)
 
         self.gridLayout.addWidget(self.label_5, 0, 0, 1, 1)
-
-        self.checkBox_plot = QCheckBox(Plot_Definer)
-        self.checkBox_plot.setObjectName(u"checkBox_plot")
-        self.checkBox_plot.setChecked(True)
-
-        self.gridLayout.addWidget(self.checkBox_plot, 4, 0, 1, 2)
 
         self.line = QFrame(Plot_Definer)
         self.line.setObjectName(u"line")
@@ -86,79 +68,89 @@ class Ui_Plot_Definer(object):
 
         self.gridLayout.addWidget(self.y_axes, 1, 0, 1, 4)
 
-        self.plotting_group = QWidget(Plot_Definer)
-        self.plotting_group.setObjectName(u"plotting_group")
-        self.gridLayout_2 = QGridLayout(self.plotting_group)
-        self.gridLayout_2.setObjectName(u"gridLayout_2")
-        self.gridLayout_2.setContentsMargins(0, 0, 0, 0)
-        self.label_2 = QLabel(self.plotting_group)
-        self.label_2.setObjectName(u"label_2")
-
-        self.gridLayout_2.addWidget(self.label_2, 1, 0, 1, 1)
-
-        self.label_3 = QLabel(self.plotting_group)
-        self.label_3.setObjectName(u"label_3")
-
-        self.gridLayout_2.addWidget(self.label_3, 3, 0, 1, 1)
-
-        self.lineEdit_ylabel = QLineEdit(self.plotting_group)
-        self.lineEdit_ylabel.setObjectName(u"lineEdit_ylabel")
-
-        self.gridLayout_2.addWidget(self.lineEdit_ylabel, 3, 1, 1, 1)
-
-        self.checkBox_xlog = QCheckBox(self.plotting_group)
-        self.checkBox_xlog.setObjectName(u"checkBox_xlog")
-
-        self.gridLayout_2.addWidget(self.checkBox_xlog, 2, 0, 1, 4)
-
-        self.checkBox_ylog = QCheckBox(self.plotting_group)
-        self.checkBox_ylog.setObjectName(u"checkBox_ylog")
-
-        self.gridLayout_2.addWidget(self.checkBox_ylog, 5, 0, 1, 2)
-
-        self.lineEdit_xlabel = QLineEdit(self.plotting_group)
-        self.lineEdit_xlabel.setObjectName(u"lineEdit_xlabel")
-
-        self.gridLayout_2.addWidget(self.lineEdit_xlabel, 1, 1, 1, 3)
-
-        self.label_4 = QLabel(self.plotting_group)
-        self.label_4.setObjectName(u"label_4")
-
-        self.gridLayout_2.addWidget(self.label_4, 3, 2, 1, 1)
-
-        self.checkBox_ylog2 = QCheckBox(self.plotting_group)
-        self.checkBox_ylog2.setObjectName(u"checkBox_ylog2")
-
[…]
