### 2024-01-17 95786ca
It is now possible to not configure instruments when starting a protocol
