### 2024-01-17 fc8ccf0
fixed problem with handing over the number of plot points
