### 2024-01-18 9e7b8e7
Add option to curate number of backuped preset files
