+++ nomad_camels/gui/settings_window.py
@@ -18,7 +18,7 @@ from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
 from PySide6.QtWidgets import (QAbstractButton, QApplication, QCheckBox, QComboBox,
     QDialog, QDialogButtonBox, QFrame, QGridLayout,
     QLabel, QLineEdit, QRadioButton, QSizePolicy,
-    QSpinBox, QWidget)
+    QSpacerItem, QSpinBox, QWidget)
 
 from nomad_camels.ui_widgets.path_button_edit import Path_Button_Edit
 
@@ -26,183 +26,302 @@ class Ui_settings_window(object):
     def setupUi(self, settings_window):
         if not settings_window.objectName():
             settings_window.setObjectName(u"settings_window")
-        settings_window.resize(391, 632)
+        settings_window.resize(634, 449)
         self.gridLayout = QGridLayout(settings_window)
         self.gridLayout.setObjectName(u"gridLayout")
-        self.line_7 = QFrame(settings_window)
-        self.line_7.setObjectName(u"line_7")
-        self.line_7.setFrameShadow(QFrame.Raised)
-        self.line_7.setLineWidth(5)
-        self.line_7.setFrameShape(QFrame.HLine)
+        self.line_vertical = QFrame(settings_window)
+        self.line_vertical.setObjectName(u"line_vertical")
+        self.line_vertical.setFrameShadow(QFrame.Raised)
+        self.line_vertical.setLineWidth(5)
+        self.line_vertical.setFrameShape(QFrame.VLine)
 
-        self.gridLayout.addWidget(self.line_7, 31, 1, 1, 3)
+        self.gridLayout.addWidget(self.line_vertical, 0, 2, 1, 1)
 
-        self.lineEdit_catalog_name = QLineEdit(settings_window)
-        self.lineEdit_catalog_name.setObjectName(u"lineEdit_catalog_name")
+        self.right = QWidget(settings_window)
+        self.right.setObjectName(u"right")
+        self.gridLayout_9 = QGridLayout(self.right)
+        self.gridLayout_9.setSpacing(9)
+        self.gridLayout_9.setObjectName(u"gridLayout_9")
+        self.gridLayout_9.setContentsMargins(3, 3, 3, 3)
+        self.eln = QWidget(self.right)
+        self.eln.setObjectName(u"eln")
+        self.gridLayout_11 = QGridLayout(self.eln)
+        self.gridLayout_11.setObjectName(u"gridLayout_11")
+        self.gridLayout_11.setContentsMargins(0, 0, 0, 0)
+        self.label_19 = QLabel(self.eln)
+        self.label_19.setObjectName(u"label_19")
 
-        self.gridLayout.addWidget(self.lineEdit_catalog_name, 22, 2, 1, 2)
+        self.gridLayout_11.addWidget(self.label_19, 1, 0, 1, 1)
 
-        self.lineEdit_repo = QLineEdit(settings_window)
-        self.lineEdit_repo.setObjectName(u"lineEdit_repo")
+        self.lineEdit_oasis = QLineEdit(self.eln)
+        self.lineEdit_oasis.setObjectName(u"lineEdit_oasis")
 
-        self.gridLayout.addWidget(self.lineEdit_repo, 25, 2, 1, 2)
+        self.gridLayout_11.addWidget(self.lineEdit_oasis, 1, 1, 1, 1)
 
-        self.label_3 = QLabel(settings_window)
-        self.label_3.setObjectName(u"label_3")
-        self.label_3.setMaximumSize(QSize(16777215, 17))
+        self.label_18 = QLabel(self.eln)
[…]

[…]
