+++ nomad_camels/gui/settings_window.py
@@ -488,6 +488,9 @@ class Ui_settings_window(object):
         self.label_10.setText(QCoreApplication.translate("settings_window", u"Databroker catalog-name", None))
         self.label_5.setText(QCoreApplication.translate("settings_window", u"Files", None))
         self.label_20.setText(QCoreApplication.translate("settings_window", u"# databroker files", None))
+#if QT_CONFIG(tooltip)
+        self.spinBox_n_databroker_files.setToolTip(QCoreApplication.translate("settings_window", u"with \"0\" all files will be kept", None))
+#endif // QT_CONFIG(tooltip)
         self.label_12.setText(QCoreApplication.translate("settings_window", u"Driver Repository URL", None))
         self.label_8.setText(QCoreApplication.translate("settings_window", u"Local drivers path", None))
         self.lineEdit_branch.setPlaceholderText(QCoreApplication.translate("settings_window", u"branch", None))
