### 2024-01-18 2c25959
Update button styles in GUI color to blue
