+++ nomad_camels/gui/device_installer.py
@@ -63,7 +63,7 @@ class Ui_Form(object):
         self.pushButton_sel_none = QPushButton(Form)
         self.pushButton_sel_none.setObjectName(u"pushButton_sel_none")
         self.pushButton_sel_none.setStyleSheet(u"QPushButton {\n"
-"        background-color: #bf8a36;\n"
+"        background-color: #A9A9A9;\n"
 "        color: white;\n"
 "        border: none;\n"
 "        padding: 2px 10px;\n"
@@ -76,7 +76,7 @@ class Ui_Form(object):
 "    }\n"
 "\n"
 "    QPushButton:hover {\n"
-"        background-color: #8a5a23;\n"
+"        background-color: #696969;\n"
 "    }")
 
         self.gridLayout.addWidget(self.pushButton_sel_none, 2, 1, 1, 1)
@@ -134,21 +134,21 @@ class Ui_Form(object):
         self.pushButton_info = QPushButton(Form)
         self.pushButton_info.setObjectName(u"pushButton_info")
         self.pushButton_info.setStyleSheet(u"QPushButton {\n"
-"        background-color: #bf8a36;\n"
-"        color: white;\n"
-"        border: none;\n"
-"        padding: 2px 10px;\n"
-"        text-align: center;\n"
-"        text-decoration: none;\n"
-"        font-size: 12px;\n"
-"        margin: 2px 2px;\n"
-"        border-radius: 6px;\n"
-"        font-weight: bold;\n"
-"    }\n"
+"    background-color: #2a4cdf; /* This is a lighter shade of blue */\n"
+"    color: white;\n"
+"    border: none;\n"
+"    padding: 2px 10px;\n"
+"    text-align: center;\n"
+"    text-decoration: none;\n"
+"    font-size: 13px;\n"
+"    margin: 2px 2px;\n"
+"    border-radius: 6px;\n"
+"    font-weight: bold;\n"
+"}\n"
 "\n"
-"    QPushButton:hover {\n"
-"        background-color: #8a5a23;\n"
-"    }")
+"QPushButton:hover {\n"
+"    background-color: #1a3cbf; /* This is a darker shade of the lighter blue for hover effect */\n"
+"}")
 
         self.gridLayout.addWidget(self.pushButton_info, 6, 2, 1, 1)
 
