+++ nomad_camels/gui/protocol_view.py
@@ -103,21 +103,21 @@ class Ui_Protocol_View(object):
         self.pushButton_move_step_up = QPushButton(self.sequence_main_widget)
         self.pushButton_move_step_up.setObjectName(u"pushButton_move_step_up")
         self.pushButton_move_step_up.setStyleSheet(u"QPushButton {\n"
-"        background-color: #bf8a36;\n"
-"        color: white;\n"
-"        border: none;\n"
-"        padding: 2px 10px;\n"
-"        text-align: center;\n"
-"        text-decoration: none;\n"
-"        font-size: 13px;\n"
-"        margin: 2px 2px;\n"
-"        border-radius: 6px;\n"
-"        font-weight: bold;\n"
-"    }\n"
+"    background-color: #2a4cdf; /* This is a lighter shade of blue */\n"
+"    color: white;\n"
+"    border: none;\n"
+"    padding: 2px 10px;\n"
+"    text-align: center;\n"
+"    text-decoration: none;\n"
+"    font-size: 13px;\n"
+"    margin: 2px 2px;\n"
+"    border-radius: 6px;\n"
+"    font-weight: bold;\n"
+"}\n"
 "\n"
-"    QPushButton:hover {\n"
-"        background-color: #8a5a23;\n"
-"    }")
+"QPushButton:hover {\n"
+"    background-color: #1a3cbf; /* This is a darker shade of the lighter blue for hover effect */\n"
+"}")
 
         self.gridLayout_8.addWidget(self.pushButton_move_step_up, 2, 0, 1, 1)
 
@@ -129,63 +129,63 @@ class Ui_Protocol_View(object):
         self.pushButton_move_step_down = QPushButton(self.sequence_main_widget)
         self.pushButton_move_step_down.setObjectName(u"pushButton_move_step_down")
         self.pushButton_move_step_down.setStyleSheet(u"QPushButton {\n"
-"        background-color: #bf8a36;\n"
-"        color: white;\n"
-"        border: none;\n"
-"        padding: 2px 10px;\n"
-"        text-align: center;\n"
-"        text-decoration: none;\n"
-"        font-size: 13px;\n"
-"        margin: 2px 2px;\n"
-"        border-radius: 6px;\n"
-"        font-weight: bold;\n"
-"    }\n"
+"    background-color: #2a4cdf; /* This is a lighter shade of blue */\n"
+"    color: white;\n"
+"    border: none;\n"
+"    padding: 2px 10px;\n"
+"    text-align: center;\n"
+"    text-decoration: none;\n"
+"    font-size: 13px;\n"
+"    margin: 2px 2px;\n"
+"    border-radius: 6px;\n"
+"    font-weight: bold;\n"
+"}\n"
 "\n"
-"    QPushButton:hover {\n"
-"        background-color: #8a5a23;\n"
-"    }")
+"QPushButton:hover {\n"
+"    background-color: #1a3cbf; /* This is a darker shade of the lighter blue for hover effect */\n"
+"}")
 
         self.gridLayout_8.addWidget(self.pushButton_move_step_down, 3, 0, 1, 1)
 
         self.pushButton_move_step_out = QPushButton(self.sequence_main_widget)
         self.pushButton_move_step_out.setObjectName(u"pushButton_move_step_out")
         self.pushButton_move_step_out.setStyleSheet(u"QPushButton {\n"
-"        background-color: #bf8a36;\n"
-"        color: white;\n"
-"        border: none;\n"
-"        padding: 2px 10px;\n"
-"        text-align: center;\n"
-"        text-decoration: none;\n"
-"        font-size: 13px;\n"
[…]
