+++ nomad_camels/gui/mainWindow_v2.py
@@ -28,7 +28,7 @@ class Ui_MainWindow(object):
     def setupUi(self, MainWindow):
         if not MainWindow.objectName():
             MainWindow.setObjectName(u"MainWindow")
-        MainWindow.resize(1060, 370)
+        MainWindow.resize(1062, 370)
         self.actionPresets = QAction(MainWindow)
         self.actionPresets.setObjectName(u"actionPresets")
         self.actionOptions = QAction(MainWindow)
@@ -95,7 +95,7 @@ class Ui_MainWindow(object):
         self.pushButton_clear_log = QPushButton(self.centralwidget)
         self.pushButton_clear_log.setObjectName(u"pushButton_clear_log")
         self.pushButton_clear_log.setStyleSheet(u"QPushButton {\n"
-"        background-color: #bf8a36;\n"
+"        background-color: #2a4cdf;\n"
 "        color: white;\n"
 "        border: none;\n"
 "        padding: 2px 10px;\n"
@@ -108,7 +108,7 @@ class Ui_MainWindow(object):
 "    }\n"
 "\n"
 "    QPushButton:hover {\n"
-"        background-color: #8a5a23;\n"
+"        background-color: #1a3cbf;\n"
 "    }")
 
         self.gridLayout_5.addWidget(self.pushButton_clear_log, 8, 9, 1, 2)
@@ -132,7 +132,10 @@ class Ui_MainWindow(object):
 "\n"
 "            QPushButton:hover {\n"
 "                background-color: #B22222;\n"
-"            }")
+"            }\n"
+"QPushButton:disabled {\n"
+"        background-color: #808080;\n"
+"    }")
 
         self.gridLayout_5.addWidget(self.pushButton_stop, 4, 11, 1, 1)
 
@@ -166,19 +169,21 @@ class Ui_MainWindow(object):
         font.setStrikeOut(False)
         self.pushButton_manage_instr.setFont(font)
         self.pushButton_manage_instr.setStyleSheet(u"QPushButton {\n"
-"                                background-color: #bf8a36; \n"
-"                                color: white; \n"
-"                                border: none; \n"
-"                                padding: 2px 4px; \n"
-"                                text-align: center; \n"
-"                                text-decoration: none; \n"
-"                                font-size: 15px; \n"
-"                                margin: 2px 2px; \n"
-"                                border-radius: 6px;\n"
-"                            }\n"
-"                            QPushButton:hover {\n"
-"                                background-color: #8a5a23;\n"
-"                            }")
+"        background-color: #2a4cdf;\n"
+"        color: white;\n"
+"        border: none;\n"
+"        padding: 2px 10px;\n"
+"        text-align: center;\n"
+"        text-decoration: none;\n"
+"        font-size: 15px;\n"
+"        margin: 2px 2px;\n"
+"        border-radius: 6px;\n"
+"        font-weight: bold;\n"
+"    }\n"
+"\n"
+"    QPushButton:hover {\n"
+"        background-color: #1a3cbf;\n"
+"    }")
 
         self.gridLayout_6.addWidget(self.pushButton_manage_instr, 0, 0, 1, 1)
 
@@ -211,20 +216,20 @@ class Ui_MainWindow(object):
         self.pushButton_editSampleInfo.setObjectName(u"pushButton_editSampleInfo")
         self.pushButton_editSampleInfo.setFont(font)
[…]

[…]
