### 2024-01-18 a18aa1c
update install button to green
