### 2024-01-22 ac5e1e9
Add extension context variables and update UI layout
