+++ nomad_camels/gui/mainWindow_v2.py
@@ -17,10 +17,10 @@ from PySide6.QtGui import (QAction, QBrush, QColor, QConicalGradient,
     QPainter, QPalette, QPixmap, QRadialGradient,
     QTransform)
 from PySide6.QtWidgets import (QApplication, QCheckBox, QComboBox, QFrame,
-    QGridLayout, QLabel, QLineEdit, QMainWindow,
-    QMenu, QMenuBar, QProgressBar, QPushButton,
-    QSizePolicy, QSpacerItem, QSplitter, QStatusBar,
-    QWidget)
+    QGridLayout, QHBoxLayout, QLabel, QLineEdit,
+    QMainWindow, QMenu, QMenuBar, QProgressBar,
+    QPushButton, QSizePolicy, QSpacerItem, QSplitter,
+    QStatusBar, QVBoxLayout, QWidget)
 
 from nomad_camels.ui_widgets.console_redirect import Console_TextEdit
 
@@ -28,7 +28,7 @@ class Ui_MainWindow(object):
     def setupUi(self, MainWindow):
         if not MainWindow.objectName():
             MainWindow.setObjectName(u"MainWindow")
-        MainWindow.resize(1021, 370)
+        MainWindow.resize(1149, 616)
         self.actionPresets = QAction(MainWindow)
         self.actionPresets.setObjectName(u"actionPresets")
         self.actionOptions = QAction(MainWindow)
@@ -142,7 +142,7 @@ class Ui_MainWindow(object):
 
         self.menu_widget = QWidget(self.centralwidget)
         self.menu_widget.setObjectName(u"menu_widget")
-        self.menu_widget.setMaximumSize(QSize(16777215, 50))
+        self.menu_widget.setMaximumSize(QSize(16777215, 60))
         self.gridLayout_6 = QGridLayout(self.menu_widget)
         self.gridLayout_6.setObjectName(u"gridLayout_6")
         self.gridLayout_6.setContentsMargins(0, 0, 0, 0)
@@ -178,105 +178,156 @@ class Ui_MainWindow(object):
 
         self.sample_widget = QWidget(self.menu_widget)
         self.sample_widget.setObjectName(u"sample_widget")
-        self.gridLayout_7 = QGridLayout(self.sample_widget)
-        self.gridLayout_7.setObjectName(u"gridLayout_7")
-        self.gridLayout_7.setContentsMargins(0, 0, 0, 0)
-        self.pushButton_editSampleInfo = QPushButton(self.sample_widget)
-        self.pushButton_editSampleInfo.setObjectName(u"pushButton_editSampleInfo")
+        self.verticalLayout_2 = QVBoxLayout(self.sample_widget)
+        self.verticalLayout_2.setSpacing(0)
+        self.verticalLayout_2.setObjectName(u"verticalLayout_2")
+        self.verticalLayout_2.setContentsMargins(0, 0, 0, 0)
+        self.sample_widget_default = QWidget(self.sample_widget)
+        self.sample_widget_default.setObjectName(u"sample_widget_default")
+        self.horizontalLayout_4 = QHBoxLayout(self.sample_widget_default)
+        self.horizontalLayout_4.setObjectName(u"horizontalLayout_4")
+        self.horizontalLayout_4.setContentsMargins(0, 3, 0, 3)
+        self.comboBox_sample = QComboBox(self.sample_widget_default)
+        self.comboBox_sample.setObjectName(u"comboBox_sample")
         font1 = QFont()
         font1.setPointSize(10)
+        self.comboBox_sample.setFont(font1)
+
+        self.horizontalLayout_4.addWidget(self.comboBox_sample)
+
+        self.pushButton_editSampleInfo = QPushButton(self.sample_widget_default)
[…]
