+++ nomad_camels/extensions/extension_contexts.py
@@ -7,3 +7,6 @@ class ELN_Context:
         self.sample_widget = ui.sample_widget
         self.session_upload_widget = ui.session_upload_widget
         self.comboBox_user_type = ui.comboBox_user_type
+        self.run_done_file_signal = ui.run_done_file_signal
+        self.extension_user = ui.extension_user
+        self.extension_sample = ui.extension_sample
