+++ nomad_camels/extensions/extension_management.py
@@ -1,12 +1,377 @@
-import sys, os
+import sys
+import os
+import requests
+import re
+import importlib
+import importlib_metadata
+import pathlib
+import pkg_resources
+import subprocess
 sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
-from nomad_camels.gui.extension_manager import Ui_Form
-from PySide6.QtWidgets import QWidget
+from nomad_camels.ui_widgets.warn_popup import WarnPopup
+from nomad_camels.utility import variables_handling
+from nomad_camels.main_classes.device_class import Simple_Config_Sub
 
-class Extension_Manager(Ui_Form, QWidget):
+from PySide6.QtWidgets import QDialog, QTableWidget, QTableWidgetItem, QLineEdit, QGridLayout, QPushButton, QCheckBox, QSplitter, QTextBrowser, QLabel, QWidget, QDialogButtonBox
+from PySide6.QtCore import Qt, QThread, Signal
+from PySide6.QtGui import QFont
+
+url = 'https://raw.githubusercontent.com/FAU-LAP/CAMELS_extensions/main/extension_list.txt'
+repo_url = 'https://raw.githubusercontent.com/FAU-LAP/CAMELS_extensions/main'
+camels_extension_regex = r'^(nomad[-_]{1}camels[-_]{1}extension[-_]{1})(.*)$'
+
+bold_font = QFont()
+bold_font.setBold(True)
+
+
+def get_online_extensions():
+    """Returns a list of all extensions."""
+    online_extensions = {}
+    try:
+        extension_str = requests.get(url).text
+        for x in extension_str.splitlines():
+            if '==' not in x:
+                continue
+            name, version = x.split('==')
+            online_extensions[name.replace('-', '_')] = version
+    except:
+        WarnPopup(None, 'Could not (completely) read extension_list.txt from repo.\n'
+                        'Check internet connectivity.',
+                          'No extensions found online')
+    return online_extensions
+
+def get_installed_extensions():
+    """Returns a list of all installed extensions."""
+    installed_extensions = {}
+    for x in importlib_metadata.distributions():
+        name = x.metadata['Name']
+        version = x.version
+        if re.match(camels_extension_regex, name, re.IGNORECASE):
+            installed_extensions[name[23:].replace('-', '_')] = version
+    return installed_extensions
+
+def get_local_extensions():
+    """Returns a list of all local extensions."""
+    local_extensions = {}
+    if 'extension_path' in variables_handling.preferences:
+        local_extension_path = variables_handling.preferences['extension_path']
+    else:
+        local_extension_path = r'C:\Users\od93yces\FAIRmat\CAMELS_extensions'
+    if not os.path.isdir(local_extension_path):
+        return local_extensions
+    sys.path.append(local_extension_path)
+    for f in pathlib.Path(local_extension_path).rglob('*'):
+        match = re.match(camels_extension_regex, f.name, re.IGNORECASE)
+        if match:
+            try:
+                sys.path.append(str(f.parent))
+                package = importlib.import_module(match.group(0))
+                name = match.group(2)
+                version = package.EXTENSION_CONFIG['version']
+                local_extensions[name] = version
+            except Exception as e:
+                print(f, e)
+    return local_extensions
+
+def get_readme_text(extension):
+    url = f'{repo_url}/{extension}/README.md'
+    return requests.get(url).text
+
+def get_license_text(extension):
+    url = f'{repo_url}/{extension}/LICENSE.txt'
+    return requests.get(url).text
+        
+
+
+class Extension_Manager(QDialog):
+    def __init__(self, preferences, parent=None):
+        super().__init__(parent=parent)
+        self.actual_preferences = preferences
+        self.preferences = dict(preferences)
[…]
