### 2024-01-24 5d46b0d
Fit values can now be displayed directly in the plot
