+++ README.md
@@ -22,6 +22,7 @@ Major Features:
 - There is now an extension feature, the first extension provides a compatibility with eLabFTW
 
 Features:
+- Fit values can now be displayed directly in the plot
 - Refactor device instantiation to be handled in another thread to keep the UI responsive, this might cause some problems with instrument-specific manual controls that are not yet updated accordingly
 - Added the plot point number to the definition already
 - It is now possible to not configure instruments when starting a protocol, this feature has to be used with caution!
