+++ README.md
@@ -27,6 +27,7 @@ Features:
 - Old databroker files may now be removed.
 - Password protection may be enabled for changing settings or presets. May be useful if the responsible person does not want users to change anything, or for lab courses.
 - There is now an extension feature, the first extension provides a compatibility with eLabFTW
+- Can now save each run in its own file.
 
 Changes:
 - the imports needed at a later time are now running in another thread after starting the main app to improve speed
