### 2024-01-25 a42e724
Add option to open save file path when right-clicking on a protocol
