+++ nomad_camels/packages.txt
@@ -1,2480 +1,0 @@
-sys
-builtins
-winreg
-marshal
-nt
-time
-zipimport
-codecs
-encodings.aliases
-encodings
-encodings.utf_8
-abc
-io
-stat
-genericpath
-ntpath
-os.path
-os
-encodings.cp1252
-types
-importlib._bootstrap
-importlib._bootstrap_external
-warnings
-importlib
-importlib._abc
-itertools
-keyword
-operator
-reprlib
-collections
-functools
-contextlib
-importlib.util
-importlib.machinery
-mpl_toolkits
-sphinxcontrib
-site
-enum
-re._constants
-re._parser
-re._casefix
-re._compiler
-copyreg
-re
-json.scanner
-json.decoder
-json.encoder
-json
-posixpath
-fnmatch
-errno
-urllib
-urllib.parse
-pathlib
-textwrap
-binascii
-zlib
-bz2
-lzma
-shutil
-struct
-threading
-zipfile
-base64
-collections.abc
-token
-tokenize
-linecache
-traceback
-math
-bisect
-random
-weakref
-tempfile
-typing.io
-typing.re
-typing
-numpy._globals
-numpy.__config__
-glob
-ctypes._endian
-ctypes
-numpy._distributor_init
-numpy._version
-numpy.version
-datetime
-numpy.core._multiarray_umath
-numpy.compat._inspect
-pickle
-numpy.compat.py3k
-numpy.compat
-numpy.core.overrides
-numpy.core.multiarray
-numpy.core.umath
-numbers
-numpy.core._string_helpers
-numpy.core._dtype
-numpy.core._type_aliases
-numpy.core.numerictypes
-numpy.core._exceptions
-contextvars
-numpy.core._ufunc_config
-numpy.core._methods
-numpy.core.fromnumeric
-numpy.core.shape_base
-numpy.core.arrayprint
-numpy.core._asarray
-numpy.core.numeric
-numpy.core.defchararray
-numpy.core.records
-numpy.core.memmap
-numpy.core.function_base
-numpy.core._machar
-numpy.core.getlimits
-numpy.core.einsumfunc
-numpy.core._multiarray_tests
-numpy.core._add_newdocs
-numpy.core._add_newdocs_scalars
-numpy.core._dtype_ctypes
-ast
-numpy.core._internal
-numpy._pytesttester
-numpy.core
-numpy.lib.mixins
-numpy.lib.ufunclike
-numpy.lib.type_check
-numpy.lib.scimath
-numpy.lib.stride_tricks
-numpy.lib.twodim_base
-numpy.linalg._umath_linalg
-numpy.linalg.linalg
-numpy.linalg
-numpy.matrixlib.defmatrix
-numpy.matrixlib
-numpy.lib.histograms
-numpy.lib.function_base
-numpy.lib.index_tricks
-numpy.lib.nanfunctions
-numpy.lib.shape_base
-numpy.lib.polynomial
-numpy.lib.utils
-numpy.lib.arraysetops
-numpy.lib.format
-numpy.lib._datasource
-numpy.lib._iotools
-numpy.lib.npyio
-numpy.lib.arrayterator
-numpy.lib.arraypad
-numpy.lib._version
-numpy.lib
-numpy.fft._pocketfft_internal
-numpy.fft._pocketfft
-numpy.fft.helper
-numpy.fft
-numpy.polynomial.polyutils
-numpy.polynomial._polybase
-numpy.polynomial.polynomial
-numpy.polynomial.chebyshev
-numpy.polynomial.legendre
-numpy.polynomial.hermite
-numpy.polynomial.hermite_e
-numpy.polynomial.laguerre
-numpy.polynomial
-cython_runtime
[…]
