### 2024-01-30 6513b94
Add signals to detect config changes in Simple_Config_Sub class
