+++ nomad_camels/utility/databroker_export.py
@@ -10,6 +10,9 @@ from datetime import datetime as dt
 import numpy as np
 import xarray
 
+from PySide6.QtWidgets import QDialog, QComboBox, QPushButton, QGridLayout, QLabel, QCheckBox
+
+sys.path.append(r'C:\Users\od93yces\FAIRmat\CAMELS')
 from nomad_camels.utility.fit_variable_renaming import replace_name
 
 
@@ -338,6 +341,174 @@ def broker_to_NX(runs, filename, plot_data=None, additional_data=None,
                 group.attrs['NX_class'] = 'NXdata'
     # print('Successfully saved protocol data!')
 
+def export_h5_to_csv_json(filename, entry_name=None, export_data=True, export_metadata=True):
+    # read the h5 file and list the entries
+    with h5py.File(filename, 'r') as file:
+        entries = list(file.keys())
+    if entry_name is not None and entry_name in entries:
+        entries = [entry_name]
+    for entry_name in entries:
+        with h5py.File(filename, 'r') as file:
+            entry = file[entry_name]
+            entry_name_non_iso = clean_filename(entry_name)
+            if export_metadata:
+                metadata = h5_group_to_dict(entry)
+                fname = f'{filename.split(".")[0]}/{entry_name_non_iso}_metadata.json'
+                if not os.path.isdir(os.path.dirname(fname)):
+                    os.makedirs(os.path.dirname(fname))
+                with open(fname, 'w', encoding='utf-8') as json_file:
+                    json.dump(metadata, json_file, cls=NumpyEncoder, indent=2)
+            if export_data:
+                if 'data' not in entry:
+                    continue
+                data = entry['data']
+                export_h5_group_to_csv(data, f'{filename.split(".")[0]}/{entry_name_non_iso}/primary.csv')
+
+
+
+class NumpyEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, np.integer):
+            return int(obj)
+        elif isinstance(obj, np.floating):
+            return float(obj)
+        elif isinstance(obj, np.ndarray):
+            return obj.tolist()
+        elif isinstance(obj, np.bool_):
+            return bool(obj)
+        else:
+            return super(NumpyEncoder, self).default(obj)
+
+def h5_group_to_dict(group):
+    """
+
+    Parameters
+    ----------
+    group :
+        
+
+    Returns
+    -------
+
+    """
+    data = {}
+    for k in group.keys():
+        if k == 'data':
+            continue
+        if isinstance(group[k], h5py.Group):
+            data[k] = h5_group_to_dict(group[k])
+        else:
+            # read dataset and handle scalar data
+            if len(group[k].shape) == 0:
+                data[k] = group[k][()]
+            else:
+                data[k] = group[k][:].tolist()
+            # if data[k] is of type bytes, convert it to string
+            if isinstance(data[k], bytes):
+                data[k] = data[k].decode('utf-8')
+    return data
+        
+def export_h5_group_to_csv(group, filename):
+    import pandas as pd
+    arrs = {}
+    if not os.path.isdir(os.path.dirname(filename)):
+        os.makedirs(os.path.dirname(filename))
+    for k in group.keys():
+        if isinstance(group[k], h5py.Group):
+            export_h5_group_to_csv(group[k], f'{filename.split(".")[0]}_{k}.csv')
+        elif isinstance(group[k], h5py.Dataset):
+            # convert the dataset to a numpy array
+            arr = group[k][:]
+            # if the array is 2D save it as a dataframe
+            if len(arr.shape) == 2:
+                df = pd.DataFrame(arr)
[…]
