+++ nomad_camels/gui/mainWindow_v2.py
@@ -83,6 +83,8 @@ class Ui_MainWindow(object):
         self.actionExport_from_databroker.setObjectName(u"actionExport_from_databroker")
         self.actionManage_Extensions = QAction(MainWindow)
         self.actionManage_Extensions.setObjectName(u"actionManage_Extensions")
+        self.actionExport_CAMELS_hdf5_to_csv_json = QAction(MainWindow)
+        self.actionExport_CAMELS_hdf5_to_csv_json.setObjectName(u"actionExport_CAMELS_hdf5_to_csv_json")
         self.centralwidget = QWidget(MainWindow)
         self.centralwidget.setObjectName(u"centralwidget")
         self.gridLayout_5 = QGridLayout(self.centralwidget)
@@ -693,7 +695,10 @@ class Ui_MainWindow(object):
         self.menuHelp.addAction(self.actionDocumentation)
         self.menuHelp.addAction(self.actionReport_Bug)
         self.menuTools.addAction(self.actionUpdate_CAMELS)
+        self.menuTools.addSeparator()
         self.menuTools.addAction(self.actionExport_from_databroker)
+        self.menuTools.addAction(self.actionExport_CAMELS_hdf5_to_csv_json)
+        self.menuTools.addSeparator()
         self.menuTools.addAction(self.action_driver_builder)
         self.menuTools.addAction(self.actionEPICS_driver_builder)
         self.menuTools.addSeparator()
@@ -732,6 +737,7 @@ class Ui_MainWindow(object):
         self.actionEPICS_driver_builder.setText(QCoreApplication.translate("MainWindow", u"EPICS-driver-builder", None))
         self.actionExport_from_databroker.setText(QCoreApplication.translate("MainWindow", u"Export from databroker", None))
         self.actionManage_Extensions.setText(QCoreApplication.translate("MainWindow", u"Managa Extensions", None))
+        self.actionExport_CAMELS_hdf5_to_csv_json.setText(QCoreApplication.translate("MainWindow", u"Export CAMELS hdf5 to csv/json", None))
         self.pushButton_clear_log.setText(QCoreApplication.translate("MainWindow", u"Clear Log", None))
         self.pushButton_stop.setText(QCoreApplication.translate("MainWindow", u"Stop", None))
         self.label_logo.setText("")
