### 2024-02-01 a786731
Add option to set only resource name in Local_VISA class
