### 2024-02-05 a2f4044
Update visa_signal.py
