### 2024-02-07 #6204 @kyleniemeyer
@lapmk I appreciate your desire for this to move forward, but please understand that JOSS relies on an all-volunteer services, and we have a waitlist because all the editors in this track have full editing loads—submissions are assigned as people become available. Fortunately, we are in the process of onboarding additional editors.
