### 2024-02-07 ce7489d
Add write_delay parameter to VISA_Signal and VISA_Signal_RO classes
