### 2024-02-07 #6204 @lapmk
Thank you very much @kyleniemeyer. Your efforts and the work of the volunteer editors of JOSS are greatly appreciated; we understand the heavy workload. Thanks for inviting @srmnitc to be editor for our manuscript. This seems indeed like a perfect fit, as he works on scientific software within MatWerk, the "neighboring" NFDI consortium to FAIRmat, in which we are developing NOMAD CAMELS presented here in our manuscript. We look forward to the review and editing process.
