### 2024-02-08 d882c3a
added advanced configuration options for change device config to be able to include variables
