### 2024-02-08 20d7c34
Refactor change_device_config.py to handle advanced configuration
