### 2024-02-08 a5072d6
Update prompt and value boxes' done flag names
