### 2024-02-12 4064e93
fixed config_changed signal of Simple_Config_Sub
