### 2024-02-14 9e42d0f
fixed bug where only one instrument of type could be used
