### 2024-02-15 ef45019
Add retry functionality to custom function signals
