+++ nomad_camels/bluesky_handling/custom_function_signal.py
@@ -16,11 +16,13 @@ class Custom_Function_Signal(Signal):
     trigger_function : callable
         Called when the Signal's `trigger` method is called.
     """
-    def __init__(self,  name, value=0., timestamp=None, parent=None, labels=None, kind='hinted', tolerance=None, rtolerance=None, metadata=None, cl=None, attr_name='', put_function=None, read_function=None, trigger_function=None):
+    def __init__(self,  name, value=0., timestamp=None, parent=None, labels=None, kind='hinted', tolerance=None, rtolerance=None, metadata=None, cl=None, attr_name='', put_function=None, read_function=None, trigger_function=None,
+                 retry_on_error=0):
         super().__init__(name=name, value=value, timestamp=timestamp, parent=parent, labels=labels, kind=kind, tolerance=tolerance, rtolerance=rtolerance, metadata=metadata, cl=cl, attr_name=attr_name)
         self.put_function = put_function
         self.read_function = read_function
         self.trigger_function = trigger_function
+        self.retry_on_error = retry_on_error
 
     def put(self, value, *, timestamp=None, force=False, metadata=None,
             **kwargs):
@@ -29,7 +31,7 @@ class Custom_Function_Signal(Signal):
         For further information see ophyd's documentation.
         """
         if self.put_function:
-            self.put_function(value)
+            retry_function(self.put_function, self.retry_on_error, value)
         super().put(value, timestamp=timestamp, force=force, metadata=metadata, **kwargs)
 
     def get(self):
@@ -38,7 +40,7 @@ class Custom_Function_Signal(Signal):
         For further information see ophyd's documentation.
         """
         if self.read_function:
-            self._readback = self.read_function()
+            self._readback = retry_function(self.read_function, self.retry_on_error)
         return super().get()
 
     def trigger(self):
@@ -47,7 +49,7 @@ class Custom_Function_Signal(Signal):
         For further information see ophyd's documentation.
         """
         if self.trigger_function:
-            self.trigger_function()
+            retry_function(self.trigger_function, self.retry_on_error)
         return super().trigger()
 
     def describe(self):
@@ -55,6 +57,17 @@ class Custom_Function_Signal(Signal):
         info = super().describe()
         info[self.name]['source'] = 'Custom Function'
         return info
+
+
+
+def retry_function(func, retries:int, *args, **kwargs):
+    excs = []
+    for i in range(retries + 1):
+        try:
+            return func(*args, **kwargs)
+        except Exception as e:
+            excs.append(e)
+    raise Exception(f'Failed to execute function {func} after {retries} retries. Last exception: {excs[-1]}')
 
 
 
@@ -70,10 +83,11 @@ class Custom_Function_SignalRO(SignalRO):
     trigger_function : callable
         Called when the Signal's `trigger` method is called.
     """
-    def __init__(self,  name, value=0., timestamp=None, parent=None, labels=None, kind='hinted', tolerance=None, rtolerance=None, metadata=None, cl=None, attr_name='', read_function=None, trigger_function=None):
+    def __init__(self,  name, value=0., timestamp=None, parent=None, labels=None, kind='hinted', tolerance=None, rtolerance=None, metadata=None, cl=None, attr_name='', read_function=None, trigger_function=None, retry_on_error=0):
         super().__init__(name=name, value=value, timestamp=timestamp, parent=parent, labels=labels, kind=kind, tolerance=tolerance, rtolerance=rtolerance, metadata=metadata, cl=cl, attr_name=attr_name)
         self.read_function = read_function
         self.trigger_function = trigger_function
+        self.retry_on_error = retry_on_error
 
     def get(self):
         """
@@ -81,7 +95,7 @@ class Custom_Function_SignalRO(SignalRO):
[…]
