+++ nomad_camels/bluesky_handling/visa_signal.py
@@ -23,7 +23,7 @@ class VISA_Signal(Signal):
                  labels=None, kind='hinted', tolerance=None, rtolerance=None,
                  metadata=None, cl=None, attr_name='',
                  write=None, parse=None, parse_return_type=None,
-                 retry_on_error=0, write_delay=0):
+                 retry_on_error=0, write_delay=0, retry_on_timeout=False):
         """
         Parameters
         ----------
@@ -59,6 +59,7 @@ class VISA_Signal(Signal):
         self.write_delay = write_delay
         self.parse = parse
         self.retry_on_error = retry_on_error
+        self.retry_on_timeout = retry_on_timeout
         if parse_return_type == 'str':
             self.parse_return_type = str
         elif parse_return_type == 'float':
@@ -110,7 +111,8 @@ class VISA_Signal(Signal):
         if self.parse is not None:
             val = retry_query_or_write(write_text, self.visa_instrument,
                                        self.retry_on_error,
-                                       write_delay=self.write_delay)
+                                       write_delay=self.write_delay,
+                                       retry_on_timeout=self.retry_on_timeout)
             if self.parse:
                 try:
                     if isinstance(self.parse, str):
@@ -127,7 +129,8 @@ class VISA_Signal(Signal):
         else:
             retry_query_or_write(write_text, self.visa_instrument,
                                  self.retry_on_error, True,
-                                 write_delay=self.write_delay)
+                                 write_delay=self.write_delay,
+                                 retry_on_timeout=self.retry_on_timeout)
         super().put(value, timestamp=timestamp, force=force, metadata=metadata, **kwargs)
 
     def describe(self):
@@ -139,7 +142,7 @@ class VISA_Signal(Signal):
 
 
 
-def retry_query_or_write(write_text, visa_instrument, retries, just_write=False, write_delay=0):
+def retry_query_or_write(write_text, visa_instrument, retries, just_write=False, write_delay=0, retry_on_timeout=False):
     excs = []
     while visa_instrument.currently_reading:
         time.sleep(0.1)
@@ -148,18 +151,28 @@ def retry_query_or_write(write_text, visa_instrument, retries, just_write=False,
         time.sleep(write_delay)
     for i in range(retries + 1):
         try:
-            val = None
-            if just_write:
-                visa_instrument.write(write_text)
+            if retry_on_timeout:
+                try:
+                    return single_query_or_write(write_text, visa_instrument, just_write)
+                except pyvisa.errors.VisaIOError as e:
+                    visa_instrument.clear()
+                    return single_query_or_write(write_text, visa_instrument, just_write)
             else:
-                val = visa_instrument.query(write_text)
-            visa_instrument.currently_reading = False
-            return val
+                return single_query_or_write(write_text, visa_instrument, just_write)
         except Exception as e:
             if i == retries:
                 print(excs)
                 raise Exception(e)
             excs.append(e)
+
+def single_query_or_write(write_text, visa_instrument, just_write=False):
+    val = None
+    if just_write:
+        visa_instrument.write(write_text)
+    else:
+        val = visa_instrument.query(write_text)
+    visa_instrument.currently_reading = False
+    return val
 
 
 class VISA_Signal_RO(SignalRO):
@@ -168,7 +181,7 @@ class VISA_Signal_RO(SignalRO):
                  kind='hinted', tolerance=None, rtolerance=None, metadata=None,
                  cl=None, attr_name='',
                  query='', parse=None, parse_return_type='float',
-                 retry_on_error=0, write_delay=0):
+                 retry_on_error=0, write_delay=0, retry_on_timeout=False):
         """
         Parameters
         ----------
[…]
