### 2024-02-15 a32aa5d
Add error_retry_function parameter to Custom_Function_Signal and Custom_Function_SignalRO, can be called if repeating on errors
