+++ nomad_camels/bluesky_handling/custom_function_signal.py
@@ -17,12 +17,13 @@ class Custom_Function_Signal(Signal):
         Called when the Signal's `trigger` method is called.
     """
     def __init__(self,  name, value=0., timestamp=None, parent=None, labels=None, kind='hinted', tolerance=None, rtolerance=None, metadata=None, cl=None, attr_name='', put_function=None, read_function=None, trigger_function=None,
-                 retry_on_error=0):
+                 retry_on_error=0, error_retry_function=None):
         super().__init__(name=name, value=value, timestamp=timestamp, parent=parent, labels=labels, kind=kind, tolerance=tolerance, rtolerance=rtolerance, metadata=metadata, cl=cl, attr_name=attr_name)
         self.put_function = put_function
         self.read_function = read_function
         self.trigger_function = trigger_function
         self.retry_on_error = retry_on_error
+        self.error_retry_function = error_retry_function
 
     def put(self, value, *, timestamp=None, force=False, metadata=None,
             **kwargs):
@@ -31,7 +32,7 @@ class Custom_Function_Signal(Signal):
         For further information see ophyd's documentation.
         """
         if self.put_function:
-            retry_function(self.put_function, self.retry_on_error, value)
+            retry_function(self.put_function, self.retry_on_error, value, error_retry_function=self.error_retry_function)
         super().put(value, timestamp=timestamp, force=force, metadata=metadata, **kwargs)
 
     def get(self):
@@ -40,7 +41,7 @@ class Custom_Function_Signal(Signal):
         For further information see ophyd's documentation.
         """
         if self.read_function:
-            self._readback = retry_function(self.read_function, self.retry_on_error)
+            self._readback = retry_function(self.read_function, self.retry_on_error, error_retry_function=self.error_retry_function)
         return super().get()
 
     def trigger(self):
@@ -49,7 +50,7 @@ class Custom_Function_Signal(Signal):
         For further information see ophyd's documentation.
         """
         if self.trigger_function:
-            retry_function(self.trigger_function, self.retry_on_error)
+            retry_function(self.trigger_function, self.retry_on_error, error_retry_function=self.error_retry_function)
         return super().trigger()
 
     def describe(self):
@@ -60,13 +61,15 @@ class Custom_Function_Signal(Signal):
 
 
 
-def retry_function(func, retries:int, *args, **kwargs):
+def retry_function(func, retries:int, *args, error_retry_function=None, **kwargs):
     excs = []
     for i in range(retries + 1):
         try:
             return func(*args, **kwargs)
         except Exception as e:
             excs.append(e)
+            if error_retry_function:
+                error_retry_function(e)
     raise Exception(f'Failed to execute function {func} after {retries} retries. Last exception: {excs[-1]}')
 
 
@@ -83,11 +86,12 @@ class Custom_Function_SignalRO(SignalRO):
     trigger_function : callable
         Called when the Signal's `trigger` method is called.
     """
-    def __init__(self,  name, value=0., timestamp=None, parent=None, labels=None, kind='hinted', tolerance=None, rtolerance=None, metadata=None, cl=None, attr_name='', read_function=None, trigger_function=None, retry_on_error=0):
+    def __init__(self,  name, value=0., timestamp=None, parent=None, labels=None, kind='hinted', tolerance=None, rtolerance=None, metadata=None, cl=None, attr_name='', read_function=None, trigger_function=None, retry_on_error=0, error_retry_function=None):
         super().__init__(name=name, value=value, timestamp=timestamp, parent=parent, labels=labels, kind=kind, tolerance=tolerance, rtolerance=rtolerance, metadata=metadata, cl=cl, attr_name=attr_name)
         self.read_function = read_function
         self.trigger_function = trigger_function
         self.retry_on_error = retry_on_error
[…]
