### 2024-02-19 05e0e6a
fixed error message "...no attribute still_running" for manual controls that did not start correctly
