### 2024-02-19 35e7406
Add starting positions to Move_Thread constructor
