### 2024-02-19 #6371 @editorialbot
[REVIEW]: NOMAD CAMELS: Configurable Application for Measurements, Experiments and Laboratory Systems
**Submitting author:** @lapmk (<a href="http://orcid.org/0000-0003-1480-9161">Michael Krieger</a>)
**Repository:** https://github.com/FAU-LAP/NOMAD-CAMELS
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** 1.0.2
