### 2024-02-19 FAU-LAP/NOMAD-CAMELS#59 @NicolasCARPi
Why two LICENSE file? (LICENSE and LICENSE.txt)
I would suggest only having one such file in the repo.

See: https://github.com/openjournals/joss-reviews/issues/6371
