### 2024-02-19 FAU-LAP/NOMAD-CAMELS#61 @NicolasCARPi
Improper error handling
While using the software, I came across several errors. It was normal for the software to give me errors because I clicked everywhere and did not follow anything.

The issues are:

* error messages are simply the python stack trace shown in red and the raw python error in a popup window. This means users are expected to understand what is wrong with this message:

![2024-02-08-212106_246x176_scrot](https://github.com/FAU-LAP/NOMAD-CAMELS/assets/3043706/fcf55f43-aea2-4c2c-b1e8-4fc30223b06a)

or this one
![2024-02-08-212334_249x192_scrot](https://github.com/FAU-LAP/NOMAD-CAMELS/assets/3043706/fa6df2f9-ad69-488b-bfb0-78cd40dd2e53)

or this one
![2024-02-08-212405_236x143_scrot](https://github.com/FAU-LAP/NOMAD-CAMELS/assets/3043706/02bc875a-3d95-4de5-8324-4d979373f64d)

![2024-02-08-212424_280x291_scrot](https://github.com/FAU-LAP/NOMAD-CAMELS/assets/3043706/c074030b-6bf6-4f66-980b-4e5ea5a60bad)

BTW, the last one gave another error and I couldn't close the app anymore, or close the windows. Had to actively kill it with `kill` command, as `^C` wasn't handled correctly. I don't remember how I triggered it  (was a few days ago) but I'll probably manage to reproduce it at some point during my review.

* the UI/UX should prevent users from creating "wrong" actions (by disabling buttons for instance, until they can get pressed without errors)
* errors must be handled gracefully with a user friendly error message, and the stack trace could still be available for debug, but the error message must not be the raw python error. You'll want to catch all exceptions and handle them properly.

--
review issue: https://github.com/openjournals/joss-reviews/issues/6371
