### 2024-02-19 FAU-LAP/NOMAD-CAMELS#62 @NicolasCARPi
Incorrect user data directory
Hello,

Running the software created a folder `NOMAD_CAMELS_data` in my `$HOME`. :angry:

Applications **must not** create random folders in the `$HOME`, but respect XDG Base Directory Specification: https://specifications.freedesktop.org/basedir-spec/basedir-spec-0.6.html

Ideally, the software would ask where user data could be saved instead of assuming the location.

review: https://github.com/openjournals/joss-reviews/issues/6371
