### 2024-02-20 b54898b
can now run from command line using "nomad-camels"
