+++ README.md
@@ -30,6 +30,7 @@ Features:
 - Can now save each run in its own file.
 - Can now open the save file path when right-clicking on a protocol.
 - Added a tool to export CAMELS hdf5 files to csv / json
+- Can now run from command line using "nomad-camels"
 
 Changes:
 - the imports needed at a later time are now running in another thread after starting the main app to improve speed
