+++ pyproject.toml
@@ -26,3 +26,6 @@ exclude = [".desertenv",".git",".idea",".gitignore"]  # exclude packages matchin
 [project.urls]
 "GitHub Page" = "https://github.com/FAU-LAP/NOMAD-CAMELS"
 "Documentation" = "https://fau-lap.github.io/NOMAD-CAMELS/"
+
+[project.gui-scripts]
+nomad-camels = "nomad_camels:start_camels"
