### 2024-02-20 ec0d94d
Delete LICENSE.txt
Fixes issue #59
