### 2024-02-20 2fa4512
Update version to 1.0.0 in pyproject.toml
