### 2024-02-20 FAU-LAP/NOMAD-CAMELS#59 @Jo-Leh
Thank you for pointing this out!
This has been fixed.
