### 2024-02-20 2dede3e
Now asking for the path, where to save measurement data when starting up CAMELS for the first time
