### 2024-02-20 FAU-LAP/NOMAD-CAMELS#62 @Jo-Leh
Sorry for cluttering your ```$HOME```.
We will add a pop-up during the first start-up to the next version.
In the meantime you can change the directory in the settings menu. I hope this serves as a solution.
