### 2024-02-20 FAU-LAP/NOMAD-CAMELS#66 @ktahar
Hard-coded paths of developer's PC?
Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/6371

I found a HDF5 in repository root: `C--Users-od93yces-NOMAD_CAMELS_data-default_user-sadf-data_2024-01-25T13-54-39.800814%2B01-00.hdf5`. What is this for?

And some paths (C:\Users\od93yces...) are hard-coded, which looks like something in a developer's PC.
I think these should be removed.

- https://github.com/FAU-LAP/NOMAD-CAMELS/blob/6b63cf1bc3e71b539f97c0db67329fe512221d4d/nomad_camels/utility/databroker_export.py#L630
- https://github.com/FAU-LAP/NOMAD-CAMELS/blob/6b63cf1bc3e71b539f97c0db67329fe512221d4d/nomad_camels/extensions/extension_management.py#L59

 Below is commented out but maybe better to remove this too.
- https://github.com/FAU-LAP/NOMAD-CAMELS/blob/6b63cf1bc3e71b539f97c0db67329fe512221d4d/nomad_camels/nomad_integration/nomad_communication.py#L240-L245
