### 2024-02-20 FAU-LAP/NOMAD-CAMELS#62 @NicolasCARPi
> We will add a pop-up during the first start-up to the next version.

Yes, with default path suggesting `~/.config/nomad-camels` (keep it lowercase).

> In the meantime you can change the directory in the settings menu

I see.

~~Two~~Three other things:

1. you have some hardcoded paths in your source code:

![2024-02-20-195132_908x482_scrot](https://github.com/FAU-LAP/NOMAD-CAMELS/assets/3043706/af257cc3-bb61-4465-9376-2c0def46d4f6)

2. Please add a "Quit" button to the "File" menu ;)

3. It might be good to address this ASAP: `nomad-camels/venv/lib/python3.11/site-packages/nomad_camels/CAMELS_start.py:107 DeprecationWarning:
Pyarrow will become a required dependency of pandas in the next major release of pandas (pandas 3.0),
(to allow more performant data types, such as the Arrow string type, and better interoperability with other libraries)
but was not found to be installed on your system.
If this would cause problems for you,
please provide us feedback at https://github.com/pandas-dev/pandas/issues/54466`
