### 2024-02-20 FAU-LAP/NOMAD-CAMELS#65 @NicolasCARPi
and they must run at each push/release on a github action. I recommand using "black" for linting.
