### 2024-02-21 FAU-LAP/NOMAD-CAMELS#65 @ktahar
> and they must run at each push/release on a github action. I recommend using "black" for linting.

I agree. These may not be the requirements for JOSS publication but I also recommend to do them.
