### 2024-02-21 2e0d596
cleaned up some remnants from testing new code
