### 2024-02-21 2fb85bc
added information how to run CAMELS after installation
