### 2024-02-21 8b1c340
startup test should now work again with new path dialog
