### 2024-02-21 FAU-LAP/NOMAD-CAMELS#66 @Jo-Leh
These paths were added with some new parts of the code.
Thanks for spotting them! I forgot to take them out. The hdf5 file as well.
All of these will be removed, currently on https://github.com/FAU-LAP/NOMAD-CAMELS/tree/development
