### 2024-02-21 f2725d8
Added "Quit" button to file menu
