+++ nomad_camels/gui/mainWindow_v2.py
@@ -85,6 +85,8 @@ class Ui_MainWindow(object):
         self.actionManage_Extensions.setObjectName(u"actionManage_Extensions")
         self.actionExport_CAMELS_hdf5_to_csv_json = QAction(MainWindow)
         self.actionExport_CAMELS_hdf5_to_csv_json.setObjectName(u"actionExport_CAMELS_hdf5_to_csv_json")
+        self.actionQuit = QAction(MainWindow)
+        self.actionQuit.setObjectName(u"actionQuit")
         self.centralwidget = QWidget(MainWindow)
         self.centralwidget.setObjectName(u"centralwidget")
         self.gridLayout_5 = QGridLayout(self.centralwidget)
@@ -662,6 +664,8 @@ class Ui_MainWindow(object):
         self.menuFile.addAction(self.actionLoad_Backup_Preset)
         self.menuFile.addSeparator()
         self.menuFile.addAction(self.actionSettings)
+        self.menuFile.addSeparator()
+        self.menuFile.addAction(self.actionQuit)
         self.menuHelp.addAction(self.actionDocumentation)
         self.menuHelp.addAction(self.actionReport_Bug)
         self.menuTools.addAction(self.actionUpdate_CAMELS)
@@ -708,6 +712,7 @@ class Ui_MainWindow(object):
         self.actionExport_from_databroker.setText(QCoreApplication.translate("MainWindow", u"Export from databroker", None))
         self.actionManage_Extensions.setText(QCoreApplication.translate("MainWindow", u"Managa Extensions", None))
         self.actionExport_CAMELS_hdf5_to_csv_json.setText(QCoreApplication.translate("MainWindow", u"Export CAMELS hdf5 to csv/json", None))
+        self.actionQuit.setText(QCoreApplication.translate("MainWindow", u"Quit", None))
         self.pushButton_clear_log.setText(QCoreApplication.translate("MainWindow", u"Clear Log", None))
         self.pushButton_stop.setText(QCoreApplication.translate("MainWindow", u"Stop", None))
         self.label_logo.setText("")
