### 2024-02-21 FAU-LAP/NOMAD-CAMELS#62 @Jo-Leh
I am actually not sure about ```~/.config/nomad-camels```. This path is where the measured data should be saved. Would the user actually look for the data there? I was thinking about ```~/.local/share/nomad_camels_data``` would this make sense? I am not too familiar with Linux, so I am quite unsure about what good paths are.

1.) Thanks for pointing it out, this is fixed together with #66
2.) This is going to be added in the next version
3.) What would you suggest to do? In the linked issue there are some good reasons against pyarrow, so we discussed some possible options:
- We just keep everything as is, until pandas 3.0 launches, then pyarrow will come with the dependencies of pandas
- We add pyarrow to our requirements although we do not actually need it and it increases the size of the software, to remove the warning
- We keep pandas below 2.2.0, all features we needed so far are included, and the warning dissappears as well. But this might cause some issues in the future with other packages relying on pandas
