+++ nomad_camels/CAMELS_start.py
@@ -4,16 +4,20 @@ from PySide6.QtWidgets import QApplication, QDialog, QProgressBar, QGridLayout, 
 from PySide6.QtCore import Qt, QCoreApplication
 from PySide6.QtGui import QPixmap, QIcon
 import os
+
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-os.environ['QT_AUTO_SCREEN_SCALE_FACTOR'] = '1'
+os.environ["QT_AUTO_SCREEN_SCALE_FACTOR"] = "1"
 
 from pkg_resources import resource_filename
+
 # Import your main application form
 # from main_window import MainWindow
+
 
 # Create a new form for the loading screen
 class LoadingScreen(QDialog):
     """ """
+
     def __init__(self):
         super().__init__()
         self.setWindowTitle("NOMAD CAMELS - Loading...")
@@ -21,15 +25,17 @@ class LoadingScreen(QDialog):
         layout = QGridLayout()
         self.setLayout(layout)
         self.setWindowFlags(Qt.CustomizeWindowHint | Qt.WindowTitleHint)
-        self.setWindowIcon(QIcon(resource_filename('nomad_camels', 'graphics/camels_icon.png')))
+        self.setWindowIcon(
+            QIcon(resource_filename("nomad_camels", "graphics/camels_icon.png"))
+        )
 
         image = QPixmap()
-        image.load(resource_filename('nomad_camels', 'graphics/CAMELS_vertical.png'))
+        image.load(resource_filename("nomad_camels", "graphics/CAMELS_vertical.png"))
         image = image.scaled(403, 308)
         image_label = QLabel()
         image_label.setPixmap(image)
 
-        self.label = QLabel('loading...')
+        self.label = QLabel("loading...")
 
         self.progress_bar = QProgressBar(self)
         layout.addWidget(image_label, 0, 0)
@@ -44,7 +50,7 @@ class LoadingScreen(QDialog):
         Parameters
         ----------
         value :
-            
+
 
         Returns
         -------
@@ -58,7 +64,7 @@ class LoadingScreen(QDialog):
         Parameters
         ----------
         text :
-            
+
 
         Returns
         -------
@@ -70,6 +76,7 @@ class LoadingScreen(QDialog):
 # Show the loading screen and import your packages
 def start_camels():
     import os
+
     appdata_path = f'{os.getenv("LOCALAPPDATA")}/nomad_camels'
     app = QCoreApplication.instance()
     if app is None:
@@ -80,21 +87,24 @@ def start_camels():
     loading_screen = LoadingScreen()
     loading_screen.show()
     import os.path
+
     file_dir = os.path.dirname(__file__)
-    package_file = f'{appdata_path}/startup_packages.txt'
+    package_file = f"{appdata_path}/startup_packages.txt"
     # package_file = f'{file_dir}/packages.txt'
     if os.path.isfile(package_file):
-        with open(package_file, 'r', encoding='utf-8') as f:
+        with open(package_file, "r", encoding="utf-8") as f:
             package_list = [x.rstrip() for x in f.readlines()]
     else:
         package_list = []
     n = len(package_list) + 1
 
     from PySide6.QtCore import QThread, Signal
+
     # Create a thread to import the packages
 
     class ImportThread(QThread):
         """ """
+
         update_progress = Signal(int)
         update_text = Signal(str)
 
@@ -103,17 +113,16 @@ def start_camels():
             # Import your packages here
             for i, package in enumerate(package_list):
                 self.update_progress.emit(int(i / n * 96))
-                self.update_text.emit(f'loading {package}...')
+                self.update_text.emit(f"loading {package}...")
                 try:
                     __import__(package)
                 except ModuleNotFoundError:
[…]
