+++ nomad_camels/gui/for_loop.py
@@ -8,44 +8,81 @@
 ## WARNING! All changes made in this file will be lost when recompiling UI file!
 ################################################################################
 
-from PySide6.QtCore import (QCoreApplication, QDate, QDateTime, QLocale,
-    QMetaObject, QObject, QPoint, QRect,
-    QSize, QTime, QUrl, Qt)
-from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
-    QFont, QFontDatabase, QGradient, QIcon,
-    QImage, QKeySequence, QLinearGradient, QPainter,
-    QPalette, QPixmap, QRadialGradient, QTransform)
-from PySide6.QtWidgets import (QApplication, QCheckBox, QComboBox, QGridLayout,
-    QHeaderView, QLabel, QLineEdit, QPushButton,
-    QSizePolicy, QSpacerItem, QTableWidget, QTableWidgetItem,
-    QWidget)
+from PySide6.QtCore import (
+    QCoreApplication,
+    QDate,
+    QDateTime,
+    QLocale,
+    QMetaObject,
+    QObject,
+    QPoint,
+    QRect,
+    QSize,
+    QTime,
+    QUrl,
+    Qt,
+)
+from PySide6.QtGui import (
+    QBrush,
+    QColor,
+    QConicalGradient,
+    QCursor,
+    QFont,
+    QFontDatabase,
+    QGradient,
+    QIcon,
+    QImage,
+    QKeySequence,
+    QLinearGradient,
+    QPainter,
+    QPalette,
+    QPixmap,
+    QRadialGradient,
+    QTransform,
+)
+from PySide6.QtWidgets import (
+    QApplication,
+    QCheckBox,
+    QComboBox,
+    QGridLayout,
+    QHeaderView,
+    QLabel,
+    QLineEdit,
+    QPushButton,
+    QSizePolicy,
+    QSpacerItem,
+    QTableWidget,
+    QTableWidgetItem,
+    QWidget,
+)
 
 from nomad_camels.ui_widgets.path_button_edit import Path_Button_Edit
 from nomad_camels.ui_widgets.variable_tool_tip_box import Variable_Box
 
+
 class Ui_for_loop_config(object):
     def setupUi(self, for_loop_config):
         if not for_loop_config.objectName():
-            for_loop_config.setObjectName(u"for_loop_config")
+            for_loop_config.setObjectName("for_loop_config")
         for_loop_config.resize(380, 476)
         self.gridLayout = QGridLayout(for_loop_config)
-        self.gridLayout.setObjectName(u"gridLayout")
+        self.gridLayout.setObjectName("gridLayout")
         self.gridLayout.setContentsMargins(0, 0, 0, 0)
         self.pushButton_add_point = QPushButton(for_loop_config)
-        self.pushButton_add_point.setObjectName(u"pushButton_add_point")
+        self.pushButton_add_point.setObjectName("pushButton_add_point")
         self.pushButton_add_point.setMinimumSize(QSize(30, 23))
         self.pushButton_add_point.setMaximumSize(QSize(30, 16777215))
 
         self.gridLayout.addWidget(self.pushButton_add_point, 1, 3, 1, 1)
 
         self.label_9 = QLabel(for_loop_config)
-        self.label_9.setObjectName(u"label_9")
+        self.label_9.setObjectName("label_9")
         self.label_9.setMaximumSize(QSize(80, 16777215))
 
         self.gridLayout.addWidget(self.label_9, 1, 2, 1, 1)
 
         self.pushButton_del_point = QPushButton(for_loop_config)
-        self.pushButton_del_point.setObjectName(u"pushButton_del_point")
+        self.pushButton_del_point.setObjectName("pushButton_del_point")
         self.pushButton_del_point.setMinimumSize(QSize(30, 23))
[…]
