+++ nomad_camels/manual_controls/set_panel/nomad_camels_driver_set_panel_device/set_panel_device.py
@@ -5,14 +5,19 @@ from nomad_camels.utility import variables_handling
 
 from .set_panel_device_ophyd import Set_Panel_Device
 
+
 class subclass(device_class.Device):
     def __init__(self, **kwargs):
-        super().__init__(name='set_panel_device', virtual=True,
-                         ophyd_device=Set_Panel_Device,
-                         ophyd_class_name='Set_Panel_Device', **kwargs)
+        super().__init__(
+            name="set_panel_device",
+            virtual=True,
+            ophyd_device=Set_Panel_Device,
+            ophyd_class_name="Set_Panel_Device",
+            **kwargs,
+        )
 
     def get_necessary_devices(self):
-        channels = self.settings['channel_names']
+        channels = self.settings["channel_names"]
         devs = []
         for chan in channels:
             devs.append(variables_handling.channels[chan].device)
@@ -20,9 +25,9 @@ class subclass(device_class.Device):
         return devs
 
     def get_channels(self):
-        if 'group_names' in self.settings:
+        if "group_names" in self.settings:
             channels = copy.deepcopy(super().get_channels())
-            n = len(self.settings['group_names'])
+            n = len(self.settings["group_names"])
             chans = {}
             for i, (nam, val) in enumerate(channels.items()):
                 if i >= n:
@@ -32,9 +37,22 @@ class subclass(device_class.Device):
         return []
 
 
-
 class subclass_config(device_class.Device_Config):
-    def __init__(self, parent=None, data='', settings_dict=None,
-                 config_dict=None, additional_info=None, **kwargs):
-        super().__init__(parent, 'set_panel_device', data, settings_dict,
-                         config_dict, additional_info, **kwargs)
+    def __init__(
+        self,
+        parent=None,
+        data="",
+        settings_dict=None,
+        config_dict=None,
+        additional_info=None,
+        **kwargs,
+    ):
+        super().__init__(
+            parent,
+            "set_panel_device",
+            data,
+            settings_dict,
+            config_dict,
+            additional_info,
+            **kwargs,
+        )

[…]
