### 2024-02-21 FAU-LAP/NOMAD-CAMELS#65 @A-D-Fuchs
We already have automated tests running for every push, see [here](https://github.com/FAU-LAP/NOMAD-CAMELS/blob/development/.github/workflows/ci.yml).
We updated all Python files using "black".
