### 2024-02-21 #6371 @lapmk
Hi @NicolasCARPi and @ktahar. First of all, thank you very much for digging so deeply into our NOMAD CAMELS code. This is extremely helpful and will lead to significant improvements. We really appreciate your input.
Just to let you know, we've released v1.0.0 yesterday afternoon which comes with several improvements and fixes. We will do our best to implement your suggestions and remarks promptly. These can immediately be viewed in the [development branch](https://github.com/FAU-LAP/NOMAD-CAMELS/tree/development). We plan to release smaller patches throughout the review process.
