### 2024-02-21 c9d2e92
changed the default configuration files path for linux and mac; now the user is asked at the first startup where to put configuration files and can set the path in the settings
