+++ nomad_camels/utility/load_save_functions.py
@@ -56,11 +56,40 @@ from nomad_camels.utility.device_handling import load_local_packages
 from nomad_camels.utility import variables_handling
 from nomad_camels.ui_widgets.warn_popup import WarnPopup
 
-appdata_path = f'{getenv("LOCALAPPDATA")}/nomad_camels'
-if not isdir(appdata_path):
-    makedirs(appdata_path)
-preset_path = f"{appdata_path}/Presets/"
-backup_path = f"{preset_path}Backup/"
+
+os_name = platform.system()
+
+if os_name == "Windows":
+    # Use the APPDATA environment variable on Windows
+    data_path = os.path.join(getenv("USERPROFILE"), "Documents", "NOMAD_CAMELS_data")
+    appdata_path = os.path.join(getenv("LOCALAPPDATA"), "nomad_camels")
+elif os_name == "Linux":
+    # Use the XDG_DATA_HOME environment variable on Linux, defaulting to ~/.local/share
+    data_path = os.path.join(
+        getenv("XDG_DATA_HOME")
+        or os.path.join(os.path.expanduser("~"), ".local", "share"),
+        "nomad_camels_data",
+    )
+    appdata_path = os.path.join(
+        getenv("XDG_CONFIG_HOME") or os.path.join(os.path.expanduser("~"), ".config"),
+        "nomad_camels",
+    )
+elif os_name == "Darwin":
+    # Use ~/Library/Application Support on MacOS
+    data_path = os.path.join(
+        os.path.expanduser("~"), "Library", "Application Support", "nomad_camels_data"
+    )
+    appdata_path = os.path.join(
+        os.path.expanduser("~"), "Library", "Application Support", "nomad_camels"
+    )
+else:
+    # Default to the home directory
+    data_path = os.path.expanduser("~", "nomad_camels_data")
+    appdata_path = os.path.expanduser("~", ".nomad_camels")
+
+
+preset_path = os.path.join(appdata_path, "Presets")
+backup_path = os.path.join(preset_path, "Backup")
 save_string_list = [QComboBox, QLineEdit, QTreeView, QListView]
 save_dict_skip = [
     QWidget,
@@ -75,26 +104,6 @@ save_dict_skip = [
 ]
 
 # Get the current operating system
-os_name = platform.system()
-
-if os_name == "Windows":
-    # Use the APPDATA environment variable on Windows
-    data_path = os.path.join(
-        os.environ.get("USERPROFILE"), "Documents", "NOMAD_CAMELS_data"
-    )
-elif os_name == "Linux":
-    # Use the XDG_DATA_HOME environment variable on Linux, defaulting to ~/.local/share
-    data_path = os.environ.get("XDG_DATA_HOME") or os.path.join(
-        os.path.expanduser("~"), ".local", "share", "nomad_camels_data"
-    )
-elif os_name == "Darwin":
-    # Use ~/Library/Application Support on MacOS
-    data_path = os.path.join(
-        os.path.expanduser("~"), "Library", "Application Support", "nomad_camels_data"
-    )
-else:
-    # Default to the home directory
-    data_path = os.path.expanduser("~", "nomad_camels_data")
 
 
 standard_pref = {
@@ -132,6 +141,46 @@ standard_pref = {
 }
 
 
+def update_config_path(path):
+    global appdata_path, preset_path, backup_path
+    appdata_path = path
+    with open(
+        os.path.join(os.path.dirname(os.path.dirname(__file__)), "config_path"), "w"
+    ) as f:
+        f.write(appdata_path)
+    preset_path = os.path.join(appdata_path, "Presets")
+    backup_path = os.path.join(preset_path, "Backup")
+
+
+def check_config_path():
+    global appdata_path
[…]
