### 2024-02-21 c4ce6c8
fixed lots of paths which were coded as strings, startup packages moved back to camels directory
