### 2024-02-21 FAU-LAP/NOMAD-CAMELS#62 @Jo-Leh
Yes, exactly.
Both paths will be asked in the next version, the folder None should then be moved to a path that the user specifies.
