### 2024-02-21 897cc76
defined the matplotlib backend to use qt, might help to fix errors of switching automatically to TK
