+++ nomad_camels/main_classes/plot_2D.py
@@ -1,6 +1,9 @@
 import threading
 import numpy as np
 
+import matplotlib
+
+matplotlib.use("Qt5Agg")
 import matplotlib.pyplot as plt
 from bluesky.callbacks.mpl_plotting import LiveScatter
 from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT
