### 2024-02-21 8d925b7
paths should now exist when needed
