+++ nomad_camels/ui_widgets/console_redirect.py
@@ -1,4 +1,5 @@
-from PySide6.QtWidgets import QTextEdit
+from PySide6.QtWidgets import QTextEdit, QWidget
+from PySide6.QtCore import Signal
 from nomad_camels.utility import variables_handling
 
 import io
@@ -9,9 +10,9 @@ import sys
 class Error_Writer(io.StringIO):
     """ """
 
-    def __init__(self, textEdit, use_old_stderr=False):
+    def __init__(self, text_signal_handler, use_old_stderr=False):
         super().__init__()
-        self.textEdit = textEdit
+        self.text_signal_handler = text_signal_handler
         self.old_stderr = sys.stderr
         self.use_old_stderr = use_old_stderr
 
@@ -32,16 +33,15 @@ class Error_Writer(io.StringIO):
         text = args[0]
         if self.use_old_stderr:
             self.old_stderr.write(text)
-        self.textEdit.setTextColor(variables_handling.get_color("strong_red"))
-        self.textEdit.append(text)
+        self.text_signal_handler.write_error_signal.emit(text)
 
 
 class Text_Writer(io.StringIO):
     """ """
 
-    def __init__(self, textEdit, use_old_stdout=False):
+    def __init__(self, textEdit, text_signal_handler, use_old_stdout=False):
         super().__init__()
-        self.textEdit = textEdit
+        self.text_signal_handler = text_signal_handler
         self.old_stdout = sys.stdout
         self.use_old_stdout = use_old_stdout
 
@@ -64,8 +64,12 @@ class Text_Writer(io.StringIO):
             self.old_stdout.write(text)
         if text == "\n":
             return
-        self.textEdit.setTextColor(variables_handling.get_color("black"))
-        self.textEdit.append(text)
+        self.text_signal_handler.write_output_signal.emit(text)
+
+
+class TextSignalHanlder(QWidget):
+    write_error_signal = Signal(str)
+    write_output_signal = Signal(str)
 
 
 class Console_TextEdit(QTextEdit):
@@ -74,5 +78,16 @@ class Console_TextEdit(QTextEdit):
     def __init__(self, parent):
         super().__init__(parent)
         use_old = not sys.executable.endswith("pythonw.exe")
-        self.text_writer = Text_Writer(self, use_old)
-        self.error_writer = Error_Writer(self, use_old)
+        self.text_handler = TextSignalHanlder()
+        self.text_handler.write_output_signal.connect(self.write_output)
+        self.text_handler.write_error_signal.connect(self.write_error)
+        self.text_writer = Text_Writer(self.text_handler, use_old)
+        self.error_writer = Error_Writer(self.text_handler, use_old)
+
+    def write_error(self, text):
+        self.setTextColor(variables_handling.get_color("strong_red"))
+        self.append(text)
+
+    def write_output(self, text):
+        self.setTextColor(variables_handling.get_color("black"))
+        self.append(text)
