### 2024-02-21 efac11b
Fix logging settings in utility module
