+++ nomad_camels/utility/logging_settings.py
@@ -26,6 +26,7 @@ def update_log_settings():
     """Reads the logging settings from the current preferences and sets them for
     the logfile-handler. Including: the level of logging, the maximum size for
     the logfile, the number of old logfiles, when one is full."""
+    global my_handler
     logfile = os.path.join(load_save_functions.appdata_path, "nomad_camels.log")
     if my_handler is None:
         my_handler = RotatingFileHandler(logfile, mode="a")
@@ -40,7 +41,7 @@ def update_log_settings():
         app_log.addHandler(my_handler)
 
     my_handler.rotation_filename = logfile
-    
+
     prefs = variables_handling.preferences
     if "log_level" in prefs and prefs["log_level"]:
         log_level = prefs["log_level"]
