### 2024-02-21 92fa2ab
Fix data type of points parameter in get_range function to int
