### 2024-02-21 34147d7
fixed issue with writing to stdout
