### 2024-02-21 fb27e63
Create log file if it doesn't exist
