### 2024-02-22 61588fd
replaced usage of pkg_resources with using importlib, preventing deprecation
