### 2024-02-22 FAU-LAP/NOMAD-CAMELS#65 @Jo-Leh
Thanks for posting the full output.
Unfortunately I cannot reproduce the error you encountered, also on GitHub Actions we never encountered that error. It might be an issue with matplotlib not knowing, which backend to use, so we now tell it to use qt. This will be added in 1.0.1 which we plan to release later today.
The warnings are quite a lot, you are right. All I could find are DeprecationWarnings or similar, coming from packages using other packages. We will keep an eye on those and set restraints on the required package versions if necessary.
