### 2024-02-22 40a55f7
added documentation on how to run tests
