### 2024-02-22 6ded31e
Update GUI to disable more buttons when not possible to use them
