+++ nomad_camels/gui/protocol_view.py
@@ -3,137 +3,98 @@
 ################################################################################
 ## Form generated from reading UI file 'protocol_view.ui'
 ##
-## Created by: Qt User Interface Compiler version 6.5.2
+## Created by: Qt User Interface Compiler version 6.5.0
 ##
 ## WARNING! All changes made in this file will be lost when recompiling UI file!
 ################################################################################
 
-from PySide6.QtCore import (
-    QCoreApplication,
-    QDate,
-    QDateTime,
-    QLocale,
-    QMetaObject,
-    QObject,
-    QPoint,
-    QRect,
-    QSize,
-    QTime,
-    QUrl,
-    Qt,
-)
-from PySide6.QtGui import (
-    QBrush,
-    QColor,
-    QConicalGradient,
-    QCursor,
-    QFont,
-    QFontDatabase,
-    QGradient,
-    QIcon,
-    QImage,
-    QKeySequence,
-    QLinearGradient,
-    QPainter,
-    QPalette,
-    QPixmap,
-    QRadialGradient,
-    QTransform,
-)
-from PySide6.QtWidgets import (
-    QApplication,
-    QFrame,
-    QGridLayout,
-    QHeaderView,
-    QLabel,
-    QPushButton,
-    QSizePolicy,
-    QSplitter,
-    QToolButton,
-    QTreeView,
-    QWidget,
-)
-
+from PySide6.QtCore import (QCoreApplication, QDate, QDateTime, QLocale,
+    QMetaObject, QObject, QPoint, QRect,
+    QSize, QTime, QUrl, Qt)
+from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
+    QFont, QFontDatabase, QGradient, QIcon,
+    QImage, QKeySequence, QLinearGradient, QPainter,
+    QPalette, QPixmap, QRadialGradient, QTransform)
+from PySide6.QtWidgets import (QApplication, QFrame, QGridLayout, QHeaderView,
+    QLabel, QPushButton, QSizePolicy, QSplitter,
+    QToolButton, QTreeView, QWidget)
 
 class Ui_Protocol_View(object):
     def setupUi(self, Protocol_View):
         if not Protocol_View.objectName():
-            Protocol_View.setObjectName("Protocol_View")
+            Protocol_View.setObjectName(u"Protocol_View")
         Protocol_View.resize(617, 522)
         self.gridLayout = QGridLayout(Protocol_View)
-        self.gridLayout.setObjectName("gridLayout")
+        self.gridLayout.setObjectName(u"gridLayout")
         self.meas_splitter = QSplitter(Protocol_View)
-        self.meas_splitter.setObjectName("meas_splitter")
+        self.meas_splitter.setObjectName(u"meas_splitter")
         self.meas_splitter.setFrameShape(QFrame.NoFrame)
         self.meas_splitter.setLineWidth(10)
         self.meas_splitter.setMidLineWidth(10)
         self.meas_splitter.setOrientation(Qt.Horizontal)
         self.sequence_main_widget = QWidget(self.meas_splitter)
-        self.sequence_main_widget.setObjectName("sequence_main_widget")
+        self.sequence_main_widget.setObjectName(u"sequence_main_widget")
         self.gridLayout_8 = QGridLayout(self.sequence_main_widget)
-        self.gridLayout_8.setObjectName("gridLayout_8")
+        self.gridLayout_8.setObjectName(u"gridLayout_8")
         self.gridLayout_8.setContentsMargins(-1, 0, -1, 0)
         self.pushButton_remove_step = QPushButton(self.sequence_main_widget)
-        self.pushButton_remove_step.setObjectName("pushButton_remove_step")
[…]
