### 2024-02-22 e254a75
changed view of exception message box
