+++ nomad_camels/utility/exception_hook.py
@@ -1,5 +1,5 @@
-from traceback import print_tb
-from PySide6.QtWidgets import QMessageBox
+from traceback import format_tb, print_tb
+from PySide6.QtWidgets import QMessageBox, QTextEdit, QPushButton, QVBoxLayout
 from PySide6.QtCore import QUrl
 
 from PySide6.QtMultimedia import QSoundEffect
@@ -30,16 +30,41 @@ class ErrorMessage(QMessageBox):
         The parent widget of this Messagebox
     """
 
-    def __init__(self, msg, info_text="", parent=None):
+    def __init__(self, exc_info, parent=None):
+        tb = format_tb(exc_info[2])
+        print_tb(exc_info[2])
+        msg = str(exc_info[0].__name__)
+        info_text = str(exc_info[1])
+        logging.exception(f"{exc_info[1]} - {tb}")
+        print(msg)
+        print(info_text)
         super().__init__(parent)
         self.setWindowTitle("ERROR")
         self.setIcon(QMessageBox.Warning)
         self.setStandardButtons(QMessageBox.Ok)
         self.setText(msg)
-        print(msg)
-        print(info_text)
         if info_text:
             self.setInformativeText(info_text)
+
+        self.more_info_button = QPushButton("more information")
+        self.more_info_button.clicked.connect(self.show_more_info)
+
+        self.text_edit = QTextEdit()
+        self.text_edit.setReadOnly(True)
+        self.text_edit.setVisible(False)
+        self.text_edit.setPlainText("".join(tb))
+
+        self.layout().addWidget(self.more_info_button, 2, 0, 1, 5)
+        self.layout().addWidget(self.text_edit, 2, 0, 1, 5)
+        self.layout().addWidget(self.button(QMessageBox.Ok), 3, 0, 1, 5)
+        self.adjustSize()
+
+    def show_more_info(self):
+        self.text_edit.setVisible(True)
+        self.more_info_button.setVisible(False)
+        self.text_edit.setMinimumHeight(500)
+        self.text_edit.setMinimumWidth(800)
+        self.adjustSize()
 
 
 def exception_hook(*exc_info):
@@ -57,15 +82,15 @@ def exception_hook(*exc_info):
         return
     elif issubclass(exc_info[0], RunEngineInterrupted):
         return
-    logging.exception(f"{exc_info[1]} - {print_tb(exc_info[2])}")
     if variables_handling.preferences["play_camel_on_error"]:
-        effect = QSoundEffect()
-        effect.setSource(
-            QUrl.fromLocalFile(
-                str(resources.files(graphics) / "Camel-Groan-2-QuickSounds.com.wav")
+        try:
+            effect = QSoundEffect()
+            effect.setSource(
+                QUrl.fromLocalFile(
+                    str(resources.files(graphics) / "Camel-Groan-2-QuickSounds.com.wav")
+                )
             )
-        )
-        effect.play()
-    ErrorMessage(
-        exc_info[0].__name__, str(exc_info[1]) + "\n" + str(print_tb(exc_info[2]))
-    ).exec()
+            effect.play()
+        except Exception as e:
+            print(e)
+    ErrorMessage(exc_info).exec()
