+++ nomad_camels/MainApp_v2.py
@@ -1428,6 +1428,9 @@ class MainWindow(Ui_MainWindow, QMainWindow):
             self.instantiate_devices_thread.start()
         except Exception as e:
             self.protocol_finished()
+            if isinstance(e, IndentationError):
+                text = "The protocol did not compile correctly, please check whether there are for example any if-statements or loops that do not have children-steps."
+                raise Exception(text).with_traceback(e.__traceback__)
             raise e
 
     def propagate_exception(self, exception):
