### 2024-02-23 f97d879
updated navigation of installation and instruments pages
