### 2024-02-23 f1fa4fb
Fix formatting and wording in installation documentation
