### 2024-02-23 FAU-LAP/NOMAD-CAMELS#65 @Jo-Leh
I am still unsure about the error you encountered, @ktahar.
It might be connected to the Qt libraries and some necessary packages for it. Some libraries are now listed at https://fau-lap.github.io/NOMAD-CAMELS/doc/installation/installation_custom_unix.html#troubleshooting

Also, we uploaded a minor update (also to pypi, v1.0.1).
