### 2024-02-26 faacb90
fixed description field. You can now use special characters and new lines
