### 2024-02-26 a8f1a1b
Added check to name string of the user
