+++ nomad_camels/MainApp_v2.py
@@ -1,5 +1,6 @@
 import sys
 import os
+import re
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 sys.path.append(os.path.dirname(__file__))
@@ -498,6 +499,12 @@ class MainWindow(Ui_MainWindow, QMainWindow):
             # changing the returned dict to dataframe and back to have a
             # dictionary that is formatted as {name: {'Name': name,...}, ...}
             dat = dialog.get_data()
+            try:
+                # Check if 'name' contains any special characters excluding whitespace
+                if re.search(r'[^\w\s]', str(dat['name'][0])):
+                    raise ValueError("Name contains special characters.\nPlease use only letters, numbers and whitespace.")
+            except ValueError as e:
+                raise ValueError("Name contains special characters.\nPlease use only letters, numbers and whitespace.")
             dat["Name2"] = dat["name"]
             data = pd.DataFrame(dat)
             data.set_index("Name2", inplace=True)
