### 2024-02-26 3ecb09f
added check to sample name. no special characters
