### 2024-02-26 3fffcf0
special characters are cought when entering them as user name or sample name
