### 2024-02-26 ae84b31
updated cam_control docs
