+++ docs/source/doc/instruments/cam_control_pylablib/cam_control_pylablib.md
@@ -1,55 +1,63 @@
 # Cam-Control by PyLabLib - Scientific Camera Control
 
 ## Install
-Install the instrument using the `Manage Instruments` button of NOMAD-CAMELS or by running 
-```
+
+Install the instrument using the `Manage Instruments` button of NOMAD-CAMELS or by running
+
+```bash
 pip install nomad-camels-driver-cam-control-pylablib
 ```
+
 in the desired Python environment.
 The most resent package can be found on [PyPI](https://pypi.org/project/nomad-camels-driver-cam-control-pylablib/).
 
 ## Basics
-The backbone of this driver is the _Cam-Control_ software by [PyLabLib](https://pylablib-cam-control.readthedocs.io/). 
-> The software must be running and set up to [allow server communication](https://pylablib-cam-control.readthedocs.io/en/latest/expanding.html#control-server).
+
+The backbone of this driver is the _Cam-Control_ software by [PyLabLib](https://pylablib-cam-control.readthedocs.io/).
+
+```{note}
+The _Cam-Control_ software must be running and set up to [allow server communication](https://pylablib-cam-control.readthedocs.io/en/latest/expanding.html#control-server).
 CAMELS then connects to the server and sends requests to the server. 
+```
 
-You must set up cam-control via its own GUI and you can then use CAMELS to actually save images with the exact settings from the GUI.
-
-
-
+You must set up _Cam-Control_ via its own GUI and you can then use CAMELS to actually save images and frames with the settings from the GUI. Some of the GUI settings can be modified with CAMELS.
 
 ## Functionality
+
 ### Configuration
+
 Under `Manage Instruments` in the `Configure Instruments`-tab you can set:
-- `Exposure time`: Exposure time of the camera. Changes value of the GUI every time a protocol with the device is run. Take care to always consider this as it will overwrite any manual setting of the exposure time at protocol startup. You can manually change the exposure time while the protocol is running but this might cause the protocol to fail as the read timing might be wrong (e.g. protocol expects an exposure time of 100 ms, but you set it manually to 10s &#8594; read timeout error.)
-- `Host IP`: Set the host IP adress of the cam-control server visible in the GUI at the bottom of the `Plugins`-tab. Format should be `xxx.xxx.xxx.xxx`
-- `Port`: Set the correct port of the server, normally `18923`, but can be changed in the `settings.cfg` file of cam-control.
+
+- `Exposure time`: Exposure time of the camera. You can check the `Overwrite exposure time` checkbox to always overwrite the currently set value with the value given in this field. The value is then set every time a protocol with the device is run. This will overwrite any manual setting of the exposure time of the GUI at protocol startup.
+- `Host IP`: Set the host IP address of the cam-control server visible in the GUI at the bottom of the `Plugins`-tab. Format should be `xxx.xxx.xxx.xxx`
+- `Port`: Set the correct port of the server, normally `18923`, but can be changed in the `settings.cfg` file of _Cam-Control_.
 - `Byte length`: Number of bytes that are read. Usually some large number (>9.000.000) should work fine. A 1920x1080 frame has `1920 * 1080 * 2 = 4.147.200` bytes.
 
-
 ### Channels
-- `get_single_frame`: Get a single frame from the buffer and handle it within Python. This allows you to save the frame with all other data to an HDF5 file. Not very fast. Time between getting frames should be around 1 second. Waits for the time specified in `Wait for frame` between requests and actually reading the frame data.
-- `get_background_frame`: Get the current background frame that is subtracted from new frames. You can only get the background frame if the background is valid. Time between getting background frames should be around 1 second. Waits for the time specified in `Wait for frame` between requests and actually reading the frame data.
+
[…]
