### 2024-02-26 62a2195
Add explanation about exposure time in cam_control_pylablib.md
