### 2024-02-27 5ba2322
strip white space at the end of names
