### 2024-02-27 6c57dc6
Add validation for special characters in y-label and y-label 2
