### 2024-02-27 8ee7d62
Add input validation for plot title and x-label
