### 2024-02-27 424ce16
Fix condition for setting maxlen in change_maxlen method
