+++ docs/source/doc/instruments/instruments.md
@@ -1,6 +1,8 @@
 # Instruments
 
 - [Agilent 34401A](agilent_34401a/agilent_34401a.md)
+- Andor Newton CCD
+- Andor Shamrock 500
 - [Cam-Control by PyLabLib](cam_control_pylablib/cam_control_pylablib.md).\
    Allows control of cameras by the following manufacturers:
   - Andor
@@ -12,21 +14,27 @@
   - Princeton Instruments
   - PhotonFocus IMAQ
   - Silicon Software
-- [Keithley 237](keithley_237/keithley_237.md)
 - [Keithley 2000](keithley_2000/keithley_2000.md)
 - Keithley 2182A
+- [Keithley 237](keithley_237/keithley_237.md)
 - Keithley 2400
 - Keithley 6221
 - [Keysight B2912A](keysight_b2912a/keysight_b2912a.md)
 - [Keysight E5270B](keysight_e5270b/keysight_e5270b.md)
+- Lakeshore F41
 - Mechnoics Stage CU30Cl
+- NI DAQ connection
 - [PID](pid/pid.md) (software implementation)
+- PI E709 piezo controller
+- SwabianInstruments TimeTagger
 - Throlabs DDR25
 - Throlabs K10CR1
 - Throlabs MFF
 - Throlabs TLPM
 - Trinamic TMCM 1110
 - [Voltcraft PPS](voltcraft_pps/voltcraft_pps.md)
+- Zaber rst240b e08
+- Zurich Instruments MFLI
 
 ```{toctree}
 :maxdepth: 2
