### 2024-02-29 1a2ee62
Fix typos
This change simply fixes typos and grammatical errors in comments and
visible UI elements.
