### 2024-02-29 FAU-LAP/NOMAD-CAMELS#70 @NicolasCARPi
Missing CONTRIBUTING.md
It is essential to have a CONTRIBUTING.md file so potential contributers can have a single source of information for specificities to this repo (such as targeting the `development` branch).

It can also set ground rules.

https://github.com/openjournals/joss-reviews/issues/6371
