### 2024-02-29 FAU-LAP/NOMAD-CAMELS#62 @NicolasCARPi
Trying the `development` branch now. It asks for folders on start. Nice job! :)

Another issue:

When trying to export to csv/json by selecting Tools > Export, the file browser should start in the data folder instead of the folder where CAMELS was started from.

Talking about this UI:
![2024-02-29-225946_627x360_scrot](https://github.com/FAU-LAP/NOMAD-CAMELS/assets/3043706/4d58d388-751f-440f-a75a-e33dde18f259)

Then, after clicking export, you don't know where the files end up. It turns out they appear in the `$HOME`. It would be better to ask for a folder here too.
