### 2024-02-29 FAU-LAP/NOMAD-CAMELS#61 @NicolasCARPi
It seems the app improved vastly on that issue, too! :clap:
