+++ CONTRIBUTING.md
@@ -1,0 +1,3 @@
+# Contributing to the NOMAD CAMELS Project
+
+See our [documentation](https://fau-lap.github.io/NOMAD-CAMELS/doc/contribute/general_process.html) for more information on how to contribute to our project.
