### 2024-03-04 bcce554
Added functionality to select path for exporting hdf5 to csv/json, fixed path if not selected
