### 2024-03-04 FAU-LAP/NOMAD-CAMELS#62 @Jo-Leh
The files were supposed to appear in the same directory as the hdf5 file. This should now be the case, additionally one can select the path if the files should go somewhere else.
