### 2024-03-04 a64ad4c
added further documentation for agilent 34401a and andor newton
