### 2024-03-04 860f623
added documentation for andor shamrock 500
