+++ docs/source/doc/instruments/instruments.md
@@ -43,8 +43,8 @@ Agilent 34401A  <agilent_34401a/agilent_34401a.md>
 Andor Newton CCD <andor_newton/andor_newton.md>
 Andor Shamrock 500 <andor_shamrock/andor_shamrock.md>
 Cam-Control by PyLabLib <cam_control_pylablib/cam_control_pylablib.md>
-Keithley 237 <keithley_237/keithley_237.md>
 Keithley 2000 <keithley_2000/keithley_2000.md>
+Keithley 237 <keithley_237/keithley_237.md>
 Keysight B2912A <keysight_b2912a/keysight_b2912a.md>
 Keysight E5270B <keysight_e5270b/keysight_e5270b.md>
 Voltcraft PPS <voltcraft_pps/voltcraft_pps.md>
