### 2024-03-04 FAU-LAP/NOMAD-CAMELS#70 @A-D-Fuchs
Added the CONTRIBUTION .md file in 9c8a7619a54de583817d65b47fd086d5736967eb and updated the [documentation website](https://fau-lap.github.io/NOMAD-CAMELS/doc/contribute/general_process.html#) for contributing. The .md file simply links to the documentation page. @NicolasCARPi is this okay?
