### 2024-03-05 #6371 @srmnitc
@ktahar @NicolasCARPi perfect, thanks a lot for your efforts. For confirmation in this thread, could you please confirm that you recommend the acceptance of this paper in JOSS?
