+++ paper.md
@@ -42,13 +42,13 @@ NOMAD CAMELS (short: CAMELS) is a configurable, open-source measurement software
 
 # Statement of need
 Research data management has piqued greater and greater interest in recent years. Today, research funding agencies demand sustainable research data strategies. The key criterion is to create research data following the FAIR principles and thereby improve world-wide data-driven research [@DFG_position_paper].
-While one ingredient, electronic lab notebooks, are an important step towards FAIR data, it is equally important to record the measurement data along FAIR principles as early as possible in the research workflow. 
+While one ingredient, electronic lab notebooks, are an important step towards FAIR data, it is equally important to record the measurement data and related metadata along FAIR principles as early as possible in the research workflow. 
 
 In experimental physics many custom-built measurement setups are controlled by very specific software written by individual researchers.
 This results in a heterogeneous landscape of software fragments for measurements written in many different languages and with often incomplete documentation, making it almost impossible for other researchers to extend existing code.
 The degree to which the stored raw data is understandable varies greatly but is often unintelligible even for researchers from the same lab.
 Important metadata such as instrument settings or the actual measurement steps performed to obtain the final raw data are rarely recorded, making it virtually impossible to reproduce experiments. Therefore, the documentation of experiments is incomplete preventing FAIR research data.
-Although there are some tools available (e.g. _SweepMe!_ [@SweepMe], _iC_ [@pernstich2012], PyMoDAQ [@PyMoDAQ]) to realise control of arbitrary measurement instruments, they are frequently not open-source or their data output is not compliant to the FAIR principles.
+Although there are some tools available (e.g. _SweepMe!_ [@SweepMe], _iC_ [@pernstich2012], PyMoDAQ [@PyMoDAQ]) to realise control of arbitrary measurement instruments, they are frequently not open-source or their data output is not compliant with the FAIR principles.
 
 ![Visualization of CAMELS functionality and workflow. CAMELS connects directly with local instruments and/or large-scale lab infrastructure running network protocols, e.g. EPICS. Customizable measurements protocols are translated into Python code and executed. The output is FAIR-compliant measurement data. \label{fig:camels_overview}](pictures/Bild_CAMELS.png){ width=80% }
 
@@ -56,7 +56,7 @@ Although there are some tools available (e.g. _SweepMe!_ [@SweepMe], _iC_ [@pern
 CAMELS is an open-source tool that automatically collects all computer-accessible experimental metadata. It features a user-friendly graphical interface that enables the creation and customization of measurements without the need for programming knowledge. By default, the data is stored in a structured HDF5 file format that closely resembles the structure of the NeXus standard [@Konnecke2015; @FAIRmatNeXusProposal].
 The final HDF5 file contains both the actual measurement data and metadata in a single file, compliant to the FAIR principles. 
 
-Moreover, CAMELS allows for direct access to the _NOMAD_ repository [@scheidgenFAIRResearchData2023], [@scheidgenNOMADDistributedWebbased2023] or its on-premise installation called _NOMAD Oasis_ enabling direct linking to electronic lab notebook (ELN) entries. The user can for example connect measurements to previous experiment workflows documented in _NOMAD_ ELNs. CAMELS can subsequently upload measurement results directly into the ELN providing a simple and stream-lined data workflow.
[…]
