### 2024-03-07 #6371 @lapmk
@srmnitc We have finished the additional author tasks.
- All author information is correct
- New release v1.0.2 created including all review-related changes
- New release is archived on Zenodo with doi 10.5281/zenodo.10793952
- Metadata of Zenodo archives matches JOSS manuscript
- License of the archive matches the software license
Note: While double checking the manuscript, we found 3 spots which we slightly rephrased for better readability. I hope that this is ok.
