### 2024-03-07 #6371 @lapmk
Hi @NicolasCARPi and @ktahar. On behalf of all co-authors I would like to thank you for your time and all your valuable comments. This is our first JOSS manuscript and we really enjoyed the open and lively review process. Your expert views and tips have undoubtedly improved our software. Thank you once again.
