### 2024-03-07 #6371 @NicolasCARPi
You're welcome @lapmk. I might come back to you to discuss ELN Consortium and RO-Crate JSON-LD ;)
