### 2024-03-07 #6371 @srmnitc
@lapmk Thanks! Some of the references are webpages, and it appears without a date in the paper. There are two possible options here: either add them as markdown links directly in the text, or add the year of access as the year for each of the citations. Could you choose either of these solutions? After that I will shortly open a pull request with some final changes, and we should ready for the last steps.
