+++ paper.bib
@@ -39,39 +39,43 @@
   title = {{{CAMELS}} - {{Configurable Application}} for {{Measurements}}, {{Experiments}} and {{Laboratory Systems}} \textemdash{} {{NOMAD-CAMELS}} Documentation},
   author = {Fuchs, Alexander D. and Lehmeyer, Johannes A. F.},
   journal = {NOMAD CAMELS - Documentation},
+  year = {2023},
   urldate = {2023-10-31},
   url = {https://fau-lap.github.io/NOMAD-CAMELS/}
 }
 
 @misc{PyMoDAQ,
   title = {PyMoDAQ},
+  year = {2023},
   urldate = {2023-10-30},
   url = {https://pymodaq.cnrs.fr/}
 }
 
 @misc{PyVISA,
   title = {PyVISA},
+  year = {2023},
   urldate = {2023-10-30},
   url = {https://pyvisa.readthedocs.io/}
 }
 
 @misc{FAIRmatNeXusProposal,
   title = {FAIRmat NeXus Proposal},
+  year = {2023},
   urldate = {2023-10-30},
   url = {https://fairmat-nfdi.github.io/nexus-fairmat-proposal/9636feecb79bb32b828b1a9804269573256d7696/fairmat-cover.html}
 }
 
 @misc{DFG_position_paper,
-  title = {{Förderung von Informationsinfrastrukturen für die Wissenschaft}},
-  shorttitle = {DFG Position Paper},
+  title = {DFG Position Paper: {Förderung von Informationsinfrastrukturen für die Wissenschaft}},
+  year = {2018},
   urldate = {2023-10-30},
-  url = {https://www.dfg.de/download/pdf/foerderung/programme/lis/positionspapier_informationsinfrastrukturen.pdf},
-  year = {2018}
+  url = {https://www.dfg.de/download/pdf/foerderung/programme/lis/positionspapier_informationsinfrastrukturen.pdf}
 }
 
 @misc{PipDocumentationV23,
   title = {Pip Documentation V23.3.1},
   shorttitle = {pip},
+  year = {2023},
   urldate = {2023-10-30},
   url = {https://pip.pypa.io/}
 }
@@ -116,6 +120,7 @@
 shorttitle = {bluesky},
 keywords = {python,software library},
 title = {{Bluesky Project}},
+year = {2023},
 url = {https://blueskyproject.io/},
 urldate = {2023-10-05}
 }
@@ -139,6 +144,7 @@ year = {2019}
 @misc{EPICS,
 shorttitle = {EPICS},
 title = {{EPICS - Experimental Physics and Industrial Control System}},
+year = {2023},
 url = {https://epics-controls.org/},
 urldate = {2023-10-05}
 }
@@ -169,6 +175,7 @@ title = {{FAU-LAP/CAMELS{\_}drivers: device implementation for CAMELS}},
 url = {https://github.com/FAU-LAP/CAMELS_drivers},
 publisher = {GitHub},
 journal = {GitHub repository},
+year = {2023},
 urldate = {2023-09-28},
 }
 
@@ -196,6 +203,7 @@ year = {2018}
 @misc{SweepMe,
 shorttitle = {SweepMe},
 title = {{SweepMe! - A multi-tool measurement software}},
+year = {2023},
 url = {https://sweep-me.net/},
 urldate = {2023-09-28},
 }
