+++ paper.bib
@@ -12,8 +12,7 @@
   doi = {10.21105/joss.05388},
   urldate = {2023-11-03},
   abstract = {Scheidgen et al., (2023). NOMAD: A distributed web-based platform for managing materials science research data. Journal of Open Source Software, 8(90), 5388, https://doi.org/10.21105/joss.05388},
-  langid = {english},
-  file = {C:\Users\yh43epyd\Zotero\storage\TPTV8IN7\Scheidgen et al. - 2023 - NOMAD A distributed web-based platform for managi.pdf}
+  langid = {english}
 }
 
 
@@ -31,14 +30,12 @@
   abstract = {Scientific research is becoming increasingly data centric, which requires more effort to manage, share, and publish data.NOMAD is a web-based platform that provides research data management (RDM) for materials-science data. In addition to core RDM functions like uploading and sharing files, NOMAD automatically extracts structured data from supported file formats, normalizes, and converts data from these formats. NOMAD provides an extendable framework for managing not just files, but structured machine-actionable harmonized and inter-operable data. This is the basis for a faceted search with domain-specific filters, a comprehensive API, structured data entry via customizable ELNs, integrated data-analysis and machine-learning tools. NOMAD is run as a free public service and can additionally be operated by research institutes. Connecting NOMAD installations through the public services will allow a federated data infrastructure to share data between research institutes and further harmonize RDM within a large research domain such as materials science.},
   copyright = {Copyright (c) 2023 Markus Scheidgen, Sebastian Br\"uckner, Sandor Brockhauser, Luca M. Ghiringhelli, Felix Dietrich, Ahmed E. Mansour, Martin Albrecht, Heiko B. Weber, Silvana Botti, Martin Aeschlimann, Claudia Draxl},
   langid = {english},
-  keywords = {electronic lab notebook,FAIR data,Materials Science,Metadata,Research Data Management},
-  file = {C:\Users\yh43epyd\Zotero\storage\TYHCS7D2\Scheidgen et al. - 2023 - FAIR Research Data With NOMAD FAIRmat's Distribut.pdf}
+  keywords = {electronic lab notebook,FAIR data,Materials Science,Metadata,Research Data Management}
 }
 
 @misc{fuchsCAMELSConfigurableApplication,
-  title = {{{CAMELS}} - {{Configurable Application}} for {{Measurements}}, {{Experiments}} and {{Laboratory Systems}} \textemdash{} {{NOMAD-CAMELS}} Documentation},
+  title = {{{CAMELS}} - {{Configurable Application}} for {{Measurements}}, {{Experiments}} and {{Laboratory Systems}} \textemdash{} {{NOMAD CAMELS}} Documentation},
   author = {Fuchs, Alexander D. and Lehmeyer, Johannes A. F.},
-  journal = {NOMAD CAMELS - Documentation},
   year = {2023},
   urldate = {2023-10-31},
   url = {https://fau-lap.github.io/NOMAD-CAMELS/}
@@ -66,7 +63,7 @@
 }
 
 @misc{DFG_position_paper,
-  title = {DFG Position Paper: {Förderung von Informationsinfrastrukturen für die Wissenschaft}},
+  title = {DFG Positionspapier: {Förderung von Informationsinfrastrukturen für die Wissenschaft}},
   year = {2018},
   urldate = {2023-10-30},
   url = {https://www.dfg.de/download/pdf/foerderung/programme/lis/positionspapier_informationsinfrastrukturen.pdf}
@@ -154,7 +151,6 @@ urldate = {2023-10-05}
[…]
