+++ paper.bib
@@ -34,7 +34,7 @@
 }
 
 @misc{fuchsCAMELSConfigurableApplication,
-  title = {{{CAMELS}} - {{Configurable Application}} for {{Measurements}}, {{Experiments}} and {{Laboratory Systems}} \textemdash{} {{NOMAD CAMELS}} Documentation},
+  title = {{NOMAD CAMELS} Documentation},
   author = {Fuchs, Alexander D. and Lehmeyer, Johannes A. F.},
   year = {2023},
   urldate = {2023-10-31},
@@ -63,14 +63,14 @@
 }
 
 @misc{DFG_position_paper,
-  title = {DFG Positionspapier: {Förderung von Informationsinfrastrukturen für die Wissenschaft}},
+  title = {{DFG Positionspapier: Förderung von Informationsinfrastrukturen für die Wissenschaft}},
   year = {2018},
   urldate = {2023-10-30},
   url = {https://www.dfg.de/download/pdf/foerderung/programme/lis/positionspapier_informationsinfrastrukturen.pdf}
 }
 
 @misc{PipDocumentationV23,
-  title = {Pip Documentation V23.3.1},
+  title = {{pip documentation V23.3.1}},
   shorttitle = {pip},
   year = {2023},
   urldate = {2023-10-30},
