+++ paper.bib
@@ -64,6 +64,7 @@
 
 @misc{DFG_position_paper,
   title = {{DFG Positionspapier: Förderung von Informationsinfrastrukturen für die Wissenschaft}},
+  shorttitle = {{DFG Positionspapier}},
   year = {2018},
   urldate = {2023-10-30},
   url = {https://www.dfg.de/download/pdf/foerderung/programme/lis/positionspapier_informationsinfrastrukturen.pdf}
@@ -166,11 +167,9 @@ year = {2022}
 }
 
 @misc{CAMELS_drivers,
-shorttitle = {CAMELS-drivers},
-title = {{FAU-LAP/CAMELS{\_}drivers: device implementation for CAMELS}},
+shorttitle = {{CAMELS-drivers}},
+title = {{CAMELS{\_}drivers: instrument implementation for CAMELS}},
 url = {https://github.com/FAU-LAP/CAMELS_drivers},
-publisher = {GitHub},
-journal = {GitHub repository},
 year = {2023},
 urldate = {2023-09-28},
 }
