+++ paper.md
@@ -56,7 +56,7 @@ Although there are some tools available (e.g. _SweepMe!_ [@SweepMe], _iC_ [@pern
 CAMELS is an open-source tool that automatically collects all computer-accessible experimental metadata. It features a user-friendly graphical interface that enables the creation and customization of measurements without the need for programming knowledge. By default, the data is stored in a structured HDF5 file format that closely resembles the structure of the NeXus standard [@Konnecke2015; @FAIRmatNeXusProposal].
 The final HDF5 file contains both the actual measurement data and metadata in a single file, compliant to the FAIR principles. 
 
-CAMELS is a stand-alone desktop application. It allows for direct access to the _NOMAD_ repository [@scheidgenFAIRResearchData2023], [@scheidgenNOMADDistributedWebbased2023] or its on-premise installation called _NOMAD Oasis_ enabling direct linking to electronic lab notebook (ELN) entries. The user can for example connect measurements to previous experiment workflows documented in _NOMAD_ ELNs. CAMELS can subsequently upload measurement results directly into the ELN providing a simple and stream-lined data workflow.
+CAMELS is a stand-alone desktop application. It allows for direct access to the _NOMAD_ repository [@scheidgenFAIRResearchData2023; @scheidgenNOMADDistributedWebbased2023] or its on-premise installation called _NOMAD Oasis_ enabling direct linking to electronic lab notebook (ELN) entries. The user can for example connect measurements to previous experiment workflows documented in _NOMAD_ ELNs. CAMELS can subsequently upload measurement results directly into the ELN providing a simple and stream-lined data workflow.
 
 CAMELS builds on _bluesky_ [@Allan2019; @bluesky] initially developed to control instruments at large-scale research institutions using EPICS [@Knott1994; @EPICS]. In CAMELS, _bluesky_ is employed to orchestrate the instrument communication either directly (e.g. via PyVISA [@PyVISA])
 or via using network protocols. Existing lab infrastructure controlled by EPICS is therefore immediately accessible. A schematic overview of the functionality of CAMELS is displayed in \autoref{fig:camels_overview}.
