### 2024-03-07 #6371 @lapmk
> @lapmk Thanks! Some of the references are webpages, and it appears without a date in the paper. There are two possible options here: either add them as markdown links directly in the text, or add the year of access as the year for each of the citations. Could you please choose either of these solutions? After that I will shortly open a pull request with some final changes, and we should ready for the last steps.

Hi @srmnitc. We've opted for the second solution with the year of access as the year for each citation. All references are updated accordingly.
