+++ paper.bib
@@ -41,13 +41,6 @@
   url = {https://fau-lap.github.io/NOMAD-CAMELS/}
 }
 
-@misc{EPICS,
-  title = {EPICS: Experimental Physics and Industrial Control System},
-  year = {2024},
-  urldate = {2024-07-03},
-  url = {https://epics-controls.org/}
-}
-
 @misc{PyMoDAQ,
   title = {PyMoDAQ},
   year = {2023},
