### 2024-03-07 #6371 @lapmk
> @lapmk I made some very minor changes to the paper for which I have opened a pull request: [FAU-LAP/NOMAD-CAMELS#72](https://github.com/FAU-LAP/NOMAD-CAMELS/pull/72)
>
> feel free to use the suggestions as needed/merge it/or implement yourself. Once you are done, please let me know and I will go ahead with next steps.

@srmnitc Done. Thanks a lot for your edits.
