### 2024-03-08 3455534
Updates JOSS paper
