### 2024-03-08 #6371 @kyleniemeyer
Hi @lapmk, I made a few edits to the paper via https://github.com/FAU-LAP/NOMAD-CAMELS/pull/73, could you review and merge?
