### 2024-03-09 #6371 @lapmk
> Hi @lapmk, I made a few edits to the paper via [FAU-LAP/NOMAD-CAMELS#73](https://github.com/FAU-LAP/NOMAD-CAMELS/pull/73), could you review and merge?

@kyleniemeyer Thank you very much for your edits. Merged. I understand that you didn't like the examples places in parentheses, however, I would prefer to keep references separate from words in the text if this is acceptable. Thus, I made two minor changes in lines 17-18 and 58 in the compiled manuscript. Is this fine with you?
