### 2024-03-11 8773f69
Fix bug preventing set panels from starting
