### 2024-03-11 3e0c274
Create CITATION.cff
